cast::base

Struct NumberWithBase

pub struct NumberWithBase {
    number: U256,
    is_nonnegative: bool,
    base: Base,
}

Utility struct for parsing numbers and formatting them into different bases.

Example

use cast::base::NumberWithBase;
use alloy_primitives::U256;

let number: NumberWithBase = U256::from(12345).into();
assert_eq!(number.format(), "12345");

// Debug uses number.base() to determine which base to format to, which defaults to Base::Decimal
assert_eq!(format!("{:?}", number), "12345");

// Display uses Base::Decimal
assert_eq!(format!("{}", number), "12345");

// The alternate formatter ("#") prepends the base's prefix
assert_eq!(format!("{:x}", number), "3039");
assert_eq!(format!("{:#x}", number), "0x3039");

assert_eq!(format!("{:b}", number), "11000000111001");
assert_eq!(format!("{:#b}", number), "0b11000000111001");

assert_eq!(format!("{:o}", number), "30071");
assert_eq!(format!("{:#o}", number), "0o30071");

Fields
number: U256

The number.

is_nonnegative: bool

Whether the number is positive or zero.

base: Base

The base to format to.

Implementations

impl NumberWithBase

pub fn new(number: impl Into<U256>, is_nonnegative: bool, base: Base) -> Self

pub fn with_base(&self, base: Base) -> Self

Creates a copy of the number with the provided base.

pub fn parse_int(s: &str, base: Option<&str>) -> Result<Self>

Parses a string slice into a signed integer. If base is None then it tries to determine base from the prefix, otherwise defaults to Decimal.

pub fn parse_uint(s: &str, base: Option<&str>) -> Result<Self>

Parses a string slice into an unsigned integer. If base is None then it tries to determine base from the prefix, otherwise defaults to Decimal.

pub fn number(&self) -> U256

Returns a copy of the underlying number as an unsigned integer. If the value is negative then the two’s complement of its absolute value will be returned.

pub fn is_nonnegative(&self) -> bool

Returns whether the underlying number is positive or zero.

pub fn base(&self) -> Base

Returns the underlying base. Defaults to Decimal.

pub const fn prefix(&self) -> &str

Returns the Rust standard prefix for the base.

pub fn set_base(&mut self, base: Base) -> &mut Self

Sets the number’s base to format to.

pub fn format(&self) -> String

Formats the number into the specified base.

Note: this method only formats the number into the base, without adding any prefixes, signs or padding. Refer to the std::fmt module documentation on how to format this number with the aforementioned properties.

fn _parse_int(s: &str, base: Base) -> Result<(U256, bool)>

fn _parse_uint(s: &str, base: Base) -> Result<U256>