cast

Module cmd

cast subcommands.

All subcommands should respect the foundry_config::Config. If a subcommand accepts values that are supported by the Config, then the subcommand should implement figment::Provider which allows the subcommand to override the config’s defaults, see [foundry_config::Config].

Modules

access_list

artifact

bind

call

constructor_args

create2

creation_code

da_estimate

Estimates the data availability size of a block for opstack.

estimate

find_block

interface

logs

mktx

rpc

run

send

storage

txpool

wallet