cast::cmd

Module interface

Structs

InterfaceArgs

CLI arguments for cast interface.

InterfaceSource 🔒

Functions

fetch_abi_from_etherscan

Fetches the ABI of a contract from Etherscan.

get_interfaces 🔒

Converts a vector of tuples containing the ABI and contract name into a vector of InterfaceSource objects.

load_abi_from_artifact 🔒

Load the ABI from the artifact of a locally compiled contract.

load_abi_from_file

Load the ABI from a file.