cast::cmd::logs

Function build_filter

fn build_filter(
    from_block: Option<BlockNumberOrTag>,
    to_block: Option<BlockNumberOrTag>,
    address: Option<Address>,
    sig_or_topic: Option<String>,
    topics_or_args: Vec<String>,
) -> Result<Filter, Error>

Builds a Filter by first trying to parse the sig_or_topic as an event signature. If successful, topics_or_args is parsed as indexed inputs and converted to topics. Otherwise, sig_or_topic is prepended to topics_or_args and used as raw topics.