Module cast::cmd::snapshot

Snapshot command

Structs

SnapshotArgs

SnapshotConfig

Additional filters that can be applied on the test results

SnapshotDiff

A Snapshot entry diff

SnapshotEntry

A general entry in a snapshot file

Enums

Format

Statics

RE_BASIC_SNAPSHOT_ENTRY

A regex that matches a basic snapshot entry like testDeposit() (gas: 58804)

Functions

check

Compares the set of tests with an existing snapshot

diff

Compare the set of tests with an existing snapshot

fmt_change

fmt_pct_change

read_snapshot

Reads a list of snapshot entries from a snapshot file

write_to_snapshot_file

Writes a series of tests to a snapshot file