cast::cmd::wallet

Module vanity

Structs

HexMatcher

Matches start and end hex.

LeftExactRightRegexMatcher

Matches start hex and end regex.

LeftHexMatcher

Matches only start hex.

LeftRegexRightExactMatcher

Matches start regex and end hex.

RegexMatcher

Matches start and end regex.

RightHexMatcher

Matches only end hex.

SingleRegexMatcher

Matches a single regex.

VanityArgs

CLI arguments for cast wallet vanity.

WalletData 🔒

WalletData contains address and private_key information for a wallet.

Wallets 🔒

Wallets is a collection of WalletData.

Traits

VanityMatcher

A trait to match vanity addresses.

Functions

create_matcher

Creates a matcher function, which takes a reference to a GeneratedWallet and returns whether it found a match or not by using matcher.

create_nonce_matcher

Creates a contract address matcher function that uses the specified nonce. The returned function takes a reference to a GeneratedWallet and returns whether the contract address created with the nonce matches using matcher.

find_vanity_address

Generates random wallets until matcher matches the wallet address, returning the wallet.

find_vanity_address_with_nonce

Generates random wallets until matcher matches the contract address created at nonce, returning the wallet.

generate_wallet

Generates a random K-256 signing key and derives its Ethereum address.

parse_pattern 🔒

save_wallet_to_file 🔒

Saves the specified wallet to a ‘vanity_addresses.json’ file at the given save_path. If the file exists, the wallet data is appended to the existing content; otherwise, a new file is created.

wallet_generator

Returns an infinite parallel iterator which yields a GeneratedWallet.

Type Aliases

GeneratedWallet

Type alias for the result of generate_wallet.