cast::cmd::wallet::vanity

Function create_nonce_matcher

pub fn create_nonce_matcher<T: VanityMatcher>(
    matcher: T,
    nonce: u64,
) -> impl Fn(&GeneratedWallet) -> bool

Creates a contract address matcher function that uses the specified nonce. The returned function takes a reference to a GeneratedWallet and returns whether the contract address created with the nonce matches using matcher.