Module cast::executors

EVM executor abstractions, which can execute calls.

Used for running tests, scripts, and interacting with the inner backend which holds the state.

Modules

ITest

Module containing a contract’s types and functions.

fuzz

invariant

Structs

CallResult

The result of a call.

DeployResult

The result of a deployment.

ExecutionErr

Represents the context after an execution error occurred.

Executor

EVM executor.

ExecutorBuilder

The builder that allows to configure an evm Executor which a stack of optional revm::Inspectors, such as Cheatcodes.

FuzzedExecutor

Wrapper around an Executor which provides fuzzing support using [proptest].

InvariantExecutor

Wrapper around any Executor implementor which provides fuzzing support using [proptest].

RawCallResult

The result of a raw call.

TracingExecutor

A default executor with tracing enabled

Enums

EvmError