Function cast::executors::invariant::assert_invariants

pub fn assert_invariants(
    invariant_contract: &InvariantContract<'_>,
    executor: &Executor,
    calldata: &[(Address, (Address, Bytes))],
    invariant_failures: &mut InvariantFailures,
    shrink_sequence: bool,
    shrink_run_limit: usize
) -> Result<Option<RawCallResult>, Report>

Given the executor state, asserts that no invariant has been broken. Otherwise, it fills the external invariant_failures.failed_invariant map and returns a generic error. Either returns the call result if successful, or nothing if there was an error.