cast::fuzz::strategies

Function fuzz_calldata_from_state

pub fn fuzz_calldata_from_state(
    func: Function,
    state: &EvmFuzzState,
    max_fuzz_int: Option<Uint<256, 4>>,
) -> impl Strategy<Value = Bytes> + use<>

Given a function and some state, it returns a strategy which generated valid calldata for the given function’s input types, based on state taken from the EVM.