cast
Module inspectors

EVM inspectors.

Modules

cheatcodes

foundry-cheatcodes

Structs

AccessListInspector

An Inspector that collects touched accounts and storage slots.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

CheatsConfig

Additional, configurable context the Cheatcodes inspector has access to

ChiselState

An inspector for Chisel

CoverageCollector

Fuzzer

An inspector that can fuzz and collect data for that effect.

InspectorData

The collected results of InspectorStack.

InspectorStack

An inspector that calls multiple inspectors in sequence.

InspectorStackBuilder

LogCollector

An inspector that collects logs during execution.

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack