Crate cast::inspectors::cheatcodes

foundry-cheatcodes

Foundry cheatcodes implementations.

Modules

spec

foundry-cheatcodes-spec

Structs

BroadcastableTransaction

Helps collecting transactions from different forks.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

CheatsConfig

Additional, configurable context the Cheatcodes inspector has access to

Context

Contains additional, test specific resources that should be kept for the duration of the test

Error

Error thrown by cheatcodes.

ScriptWallets

Clonable wrapper around [ScriptWalletsInner].

Enums

ErrorKind

Kind of cheatcode errors.

Traits

CheatcodeDef

Cheatcode definition trait. Implemented by all [Vm] functions.

Type Aliases

BroadcastableTransactions

List of transactions that can be broadcasted.

ExpectedCallTracker

Tracks the expected calls per address.

Result

Cheatcode result type.