Struct cast::inspectors::Cheatcodes

pub struct Cheatcodes {

    pub block: Option<BlockEnv>,
    pub gas_price: Option<Uint<256, 4>>,
    pub labels: HashMap<Address, String>,
    pub prank: Option<Prank>,
    pub expected_revert: Option<ExpectedRevert>,
    pub fork_revert_diagnostic: Option<RevertDiagnostic>,
    pub accesses: Option<RecordAccess>,
    pub recorded_account_diffs_stack: Option<Vec<Vec<AccountAccess>>>,
    pub recorded_logs: Option<Vec<Log>>,
    pub last_call_gas: Option<Gas>,
    pub mocked_calls: HashMap<Address, BTreeMap<MockCallDataContext, MockCallReturnData>>,
    pub expected_calls: HashMap<Address, HashMap<Bytes, (ExpectedCallData, u64)>>,
    pub expected_emits: VecDeque<ExpectedEmit>,
    pub allowed_mem_writes: HashMap<u64, Vec<Range<u64>>, FxBuildHasher>,
    pub broadcast: Option<Broadcast>,
    pub broadcastable_transactions: VecDeque<BroadcastableTransaction>,
    pub config: Arc<CheatsConfig>,
    pub context: Context,
    pub fs_commit: bool,
    pub serialized_jsons: BTreeMap<String, BTreeMap<String, Value>>,
    pub eth_deals: Vec<DealRecord>,
    pub gas_metering: Option<Option<Gas>>,
    pub gas_metering_create: Option<Option<Gas>>,
    pub mapping_slots: Option<HashMap<Address, MappingSlots>>,
    pub pc: usize,
    pub breakpoints: HashMap<char, (Address, usize), FxBuildHasher>,

}

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

Cheatcodes can be called by contracts during execution to modify the VM environment, such as mocking addresses, signatures and altering call reverts.

Executing cheatcodes can be very powerful. Most cheatcodes are limited to evm internals, but there are also cheatcodes like ffi which can execute arbitrary commands or writeFile and readFile which can manipulate files of the filesystem. Therefore, several restrictions are implemented for these cheatcodes:

ffi, and file cheatcodes are always opt-in (via foundry config) and never enabled by default: all respective cheatcode handlers implement the appropriate checks

File cheatcodes require explicit permissions which paths are allowed for which operation, see Config.fs_permission

Only permitted accounts are allowed to execute cheatcodes in forking mode, this ensures no contract deployed on the live network is able to execute cheatcodes by simply calling the cheatcode address: by default, the caller, test contract and newly deployed contracts are allowed to execute cheatcodes

Fields
block: Option<BlockEnv>

The block environment

Used in the cheatcode handler to overwrite the block environment separately from the execution block environment.

gas_price: Option<Uint<256, 4>>

The gas price

Used in the cheatcode handler to overwrite the gas price separately from the gas price in the execution environment.

labels: HashMap<Address, String>

Address labels

prank: Option<Prank>

Prank information

expected_revert: Option<ExpectedRevert>

Expected revert information

fork_revert_diagnostic: Option<RevertDiagnostic>

Additional diagnostic for reverts

accesses: Option<RecordAccess>

Recorded storage reads and writes

recorded_account_diffs_stack: Option<Vec<Vec<AccountAccess>>>

Recorded account accesses (calls, creates) organized by relative call depth, where the topmost vector corresponds to accesses at the depth at which account access recording began. Each vector in the matrix represents a list of accesses at a specific call depth. Once that call context has ended, the last vector is removed from the matrix and merged into the previous vector.

recorded_logs: Option<Vec<Log>>

Recorded logs

last_call_gas: Option<Gas>

Cache of the amount of gas used in previous call. This is used by the lastCallGas cheatcode.

mocked_calls: HashMap<Address, BTreeMap<MockCallDataContext, MockCallReturnData>>

Mocked calls

expected_calls: HashMap<Address, HashMap<Bytes, (ExpectedCallData, u64)>>

Expected calls

expected_emits: VecDeque<ExpectedEmit>

Expected emits

allowed_mem_writes: HashMap<u64, Vec<Range<u64>>, FxBuildHasher>

Map of context depths to memory offset ranges that may be written to within the call depth.

broadcast: Option<Broadcast>

Current broadcasting information

broadcastable_transactions: VecDeque<BroadcastableTransaction>

Scripting based transactions

config: Arc<CheatsConfig>

Additional, user configurable context this Inspector has access to when inspecting a call

context: Context

Test-scoped context holding data that needs to be reset every test run

fs_commit: bool

Whether to commit FS changes such as file creations, writes and deletes. Used to prevent duplicate changes file executing non-committing calls.

serialized_jsons: BTreeMap<String, BTreeMap<String, Value>>

Serialized JSON values.

eth_deals: Vec<DealRecord>

All recorded ETH deals.

gas_metering: Option<Option<Gas>>

Holds the stored gas info for when we pause gas metering. It is an Option<Option<..>> because the call callback in an Inspector doesn’t get access to the revm::Interpreter which holds the revm::Gas struct that we need to copy. So we convert it to a Some(None) in apply_cheatcode, and once we have the interpreter, we copy the gas struct. Then each time there is an execution of an operation, we reset the gas.

gas_metering_create: Option<Option<Gas>>

Holds stored gas info for when we pause gas metering, and we’re entering/inside CREATE / CREATE2 frames. This is needed to make gas meter pausing work correctly when paused and creating new contracts.

mapping_slots: Option<HashMap<Address, MappingSlots>>

Mapping slots.

pc: usize

The current program counter.

breakpoints: HashMap<char, (Address, usize), FxBuildHasher>

Breakpoints supplied by the breakpoint cheatcode. char -> (address, pc)

Implementations

impl Cheatcodes

pub fn new(config: Arc<CheatsConfig>) -> Cheatcodes

Creates a new Cheatcodes with the given settings.

pub fn script_wallets(&self) -> Option<&ScriptWallets>

Returns the configured script wallets.

pub fn on_revert<DB>(&mut self, ecx: &mut EvmContext<DB>)
where DB: DatabaseExt,

Called when there was a revert.

Cleanup any previously applied cheatcodes that altered the state in such a way that revm’s revert would run into issues.

pub fn call_with_executor<DB>( &mut self, ecx: &mut EvmContext<DB>, call: &mut CallInputs, executor: &mut impl CheatcodesExecutor, ) -> Option<CallOutcome>
where DB: DatabaseExt,