cast::revm::context

Module journal

This module contains Journal struct and implements JournalTr trait for it.

Entry submodule contains JournalEntry and JournalEntryTr traits. and inner submodule contains JournalInner struct that contains state.

Modules

entry

Contains the journal entry trait and implementations.

inner

Module containing the JournalInner that is part of crate::Journal.

warm_addresses

This module contains WarmAddresses struct that stores addresses that are warm loaded.

Structs

Journal

A journal of state changes internal to the EVM

JournalInner

Inner journal state that contains journal and state changes.

Enums

JournalEntry

Journal entries that are used to track changes to the state and are used to revert it.

Traits

JournalEntryTr

Trait for tracking and reverting state changes in the EVM. Journal entry contains information about state changes that can be reverted.