cast::revm::context

Module result

Result of the EVM execution. Containing both execution result, state and errors.

ExecutionResult is the result of the EVM execution.

InvalidTransaction is the error that is returned when the transaction is invalid.

InvalidHeader is the error that is returned when the header is invalid.

SuccessReason is the reason that the transaction successfully completed.

Structs

ExecResultAndState

Tuple containing evm execution result and state.s

Enums

EVMError

Main EVM error

ExecutionResult

Result of a transaction execution

HaltReason

Indicates that the EVM has experienced an exceptional halt.

InvalidHeader

Errors related to misconfiguration of a crate::Block.

InvalidTransaction

Transaction validation error.

OutOfGasError

Out of gas errors.

Output

Output of a transaction execution

SuccessReason

Reason a transaction successfully completed.

Traits

FromStringError

Trait for converting a string to an EVMError::Custom error.

HaltReasonTr

Trait for the halt reason.

Type Aliases

ResultAndState

Type alias for backwards compatibility.

ResultVecAndState

Tuple containing multiple execution results and state.