cast::revm::context

Struct Context

pub struct Context<BLOCK = BlockEnv, TX = TxEnv, CFG = CfgEnv, DB = EmptyDBTyped<Infallible>, JOURNAL = Journal<DB>, CHAIN = (), LOCAL = LocalContext>
where
    DB: Database,
    JOURNAL: JournalTr<Database = DB>,
    LOCAL: LocalContextTr,
{
    pub block: BLOCK,
    pub tx: TX,
    pub cfg: CFG,
    pub journaled_state: JOURNAL,
    pub chain: CHAIN,
    pub local: LOCAL,
    pub error: Result<(), ContextError<<DB as Database>::Error>>,
}

EVM context contains data that EVM needs for execution.

Fields
block: BLOCK

Block information.

tx: TX

Transaction information.

cfg: CFG

Configurations.

journaled_state: JOURNAL

EVM State with journaling support and database.

chain: CHAIN

Inner context.

local: LOCAL

Local context that is filled by execution.

error: Result<(), ContextError<<DB as Database>::Error>>

Error that happened during execution.

Implementations

impl<BLOCK, TX, DB, JOURNAL, CHAIN, LOCAL> Context<BLOCK, TX, CfgEnv, DB, JOURNAL, CHAIN, LOCAL>
where BLOCK: Block + Default, TX: Transaction + Default, DB: Database, JOURNAL: JournalTr<Database = DB>, CHAIN: Default, LOCAL: LocalContextTr + Default,

pub fn new( db: DB, spec: SpecId, ) -> Context<BLOCK, TX, CfgEnv, DB, JOURNAL, CHAIN, LOCAL>

Creates a new context with a new database type.

This will create a new Journal object.

impl<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where BLOCK: Block, TX: Transaction, CFG: Cfg, DB: Database, JOURNAL: JournalTr<Database = DB>, LOCAL: LocalContextTr,

pub fn with_new_journal<OJOURNAL>( self, journal: OJOURNAL, ) -> Context<BLOCK, TX, CFG, DB, OJOURNAL, CHAIN, LOCAL>
where OJOURNAL: JournalTr<Database = DB>,

Creates a new context with a new journal type. New journal needs to have the same database type.

pub fn with_db<ODB>( self, db: ODB, ) -> Context<BLOCK, TX, CFG, ODB, Journal<ODB>, CHAIN, LOCAL>
where ODB: Database,

Creates a new context with a new database type.

This will create a new Journal object.

pub fn with_ref_db<ODB>( self, db: ODB, ) -> Context<BLOCK, TX, CFG, WrapDatabaseRef<ODB>, Journal<WrapDatabaseRef<ODB>>, CHAIN, LOCAL>
where ODB: DatabaseRef,

Creates a new context with a new DatabaseRef type.

pub fn with_block<OB>( self, block: OB, ) -> Context<OB, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where OB: Block,

Creates a new context with a new block type.

pub fn with_tx<OTX>( self, tx: OTX, ) -> Context<BLOCK, OTX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where OTX: Transaction,

Creates a new context with a new transaction type.

pub fn with_chain<OC>( self, chain: OC, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, OC, LOCAL>

Creates a new context with a new chain type.

pub fn with_cfg<OCFG>( self, cfg: OCFG, ) -> Context<BLOCK, TX, OCFG, DB, JOURNAL, CHAIN, LOCAL>
where OCFG: Cfg,

Creates a new context with a new chain type.

pub fn with_local<OL>( self, local: OL, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, OL>
where OL: LocalContextTr,

Creates a new context with a new local context type.

pub fn modify_cfg_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut CFG),

Modifies the context configuration.

pub fn modify_block_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut BLOCK),

Modifies the context block.

pub fn modify_tx_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut TX),

Modifies the context transaction.

pub fn modify_chain_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut CHAIN),

Modifies the context chain.

pub fn modify_db_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut DB),

Modifies the context database.

pub fn modify_journal_chained<F>( self, f: F, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN, LOCAL>
where F: FnOnce(&mut JOURNAL),

Modifies the context journal.

pub fn modify_block<F>(&mut self, f: F)
where F: FnOnce(&mut BLOCK),

Modifies the context block.

pub fn modify_tx<F>(&mut self, f: F)
where F: FnOnce(&mut TX),

Modifies the context transaction.

pub fn modify_cfg<F>(&mut self, f: F)
where F: FnOnce(&mut CFG),

Modifies the context configuration.

pub fn modify_chain<F>(&mut self, f: F)
where F: FnOnce(&mut CHAIN),

Modifies the context chain.

pub fn modify_db<F>(&mut self, f: F)
where F: FnOnce(&mut DB),

Modifies the context database.

pub fn modify_journal<F>(&mut self, f: F)
where F: FnOnce(&mut JOURNAL),

Modifies the context journal.

pub fn modify_local<F>(&mut self, f: F)
where F: FnOnce(&mut LOCAL),

Modifies the local context.