cast::revm::context

Module transaction

Transaction trait Transaction and associated types.

Modules

eip2930

EIP-2930 Access list transaction interface.

eip7702

EIP-7702 Account Abstraction transaction interface.

transaction_type

Transaction type enum.

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

Authorization

An unsigned EIP-7702 authorization.

RecoveredAuthorization

A recovered authorization.

SignedAuthorization

A signed EIP-7702 authorization.

Enums

RecoveredAuthority

Represents the outcome of an attempt to recover the authority from an authorization. It can either be valid (containing an Address) or invalid (indicating recovery failure).

TransactionType

Transaction types of all Ethereum transactions

Traits

AccessListItemTr

Access list type is introduced in EIP-2930, and every transaction after it contains access list.

AuthorizationTr

Authorization trait.

Transaction

Main Transaction trait that abstracts and specifies all transaction currently supported by Ethereum

TransactionError

Transaction validity error types.