cast::revm::context::transaction::eip7702

Trait AuthorizationTr

pub trait AuthorizationTr {
    // Required methods
    fn authority(&self) -> Option<Address>;
    fn chain_id(&self) -> Uint<256, 4>;
    fn nonce(&self) -> u64;
    fn address(&self) -> Address;
}

Authorization trait.

Required Methods

fn authority(&self) -> Option<Address>

Authority address.

Note

Authority signature can be invalid, so this method returns None if the authority could not be recovered.

Valid signature Parity should be 0 or 1 and signature s-value should be less than SECP256K1N_HALF.

fn chain_id(&self) -> Uint<256, 4>

Returns the chain id from the authorization.

fn nonce(&self) -> u64

Returns the nonce.

Note

If nonce is not same as the nonce of the signer account, the authorization is skipped.

fn address(&self) -> Address

Returns the address that this account is delegated to.