cast::revm

Crate context_interface

Optimism-specific constants, types, and helpers.

Modules

block

cfg

context

local

Enums

CreateScheme

Create scheme

TransactionType

Transaction types of all Ethereum transaction

Traits

Block

Trait for retrieving block information required for execution.

Cfg

ContextSetters

ContextTr

Trait that defines the context of the EVM execution.

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

JournalTr

Trait that contains database and journal of all changes that were made to the state.

LocalContextTr

Local context used for caching initcode from Initcode transactions.

Transaction

Main Transaction trait that abstracts and specifies all transaction currently supported by Ethereum

Type Aliases

TransactTo

Transaction destination