cast::revm::context_interface

Module block

Block related types and functions.

Block trait is used to retrieve block information required for execution.

Modules

blob

Blob (EIP-4844) related functions and types. BlobExcessGasAndPrice is struct that helps with calculating blob gas price and excess blob gas.

Structs

BlobExcessGasAndPrice

Structure holding block blob excess gas and it calculates blob fee

Traits

Block

Trait for retrieving block information required for execution.

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas. Uses calc_excess_blob_gas_osaka internally.

calc_excess_blob_gas_osaka

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.