cast::revm::context_interface::block

Trait Block

pub trait Block {
    // Required methods
    fn number(&self) -> Uint<256, 4>;
    fn beneficiary(&self) -> Address;
    fn timestamp(&self) -> Uint<256, 4>;
    fn gas_limit(&self) -> u64;
    fn basefee(&self) -> u64;
    fn difficulty(&self) -> Uint<256, 4>;
    fn prevrandao(&self) -> Option<FixedBytes<32>>;
    fn blob_excess_gas_and_price(&self) -> Option<BlobExcessGasAndPrice>;

    // Provided methods
    fn blob_gasprice(&self) -> Option<u128> { ... }
    fn blob_excess_gas(&self) -> Option<u64> { ... }
}

Trait for retrieving block information required for execution.

Required Methods

fn number(&self) -> Uint<256, 4>

The number of ancestor blocks of this block (block height).

fn beneficiary(&self) -> Address

Beneficiary (Coinbase, miner) is an address that has signed the block.

This is the receiver address of priority gas rewards.

fn timestamp(&self) -> Uint<256, 4>

The timestamp of the block in seconds since the UNIX epoch.

fn gas_limit(&self) -> u64

The gas limit of the block.

fn basefee(&self) -> u64

The base fee per gas, added in the London upgrade with EIP-1559.

fn difficulty(&self) -> Uint<256, 4>

The difficulty of the block.

Unused after the Paris (AKA the merge) upgrade, and replaced by prevrandao.

fn prevrandao(&self) -> Option<FixedBytes<32>>

The output of the randomness beacon provided by the beacon chain.

Replaces difficulty after the Paris (AKA the merge) upgrade with EIP-4399.

Note: prevrandao can be found in a block in place of mix_hash.

fn blob_excess_gas_and_price(&self) -> Option<BlobExcessGasAndPrice>

Excess blob gas and blob gasprice. See also calc_excess_blob_gas and calc_blob_gasprice.

Incorporated as part of the Cancun upgrade via EIP-4844.

Provided Methods

fn blob_gasprice(&self) -> Option<u128>

See EIP-4844 and calc_blob_gasprice.

Returns None if Cancun is not enabled.

fn blob_excess_gas(&self) -> Option<u64>

Return blob_excess_gas header field. See EIP-4844.

Returns None if Cancun is not enabled.