cast::revm::context_interface::cfg

Trait Cfg

pub trait Cfg {
    type Spec: Into<SpecId> + Clone;

    // Required methods
    fn chain_id(&self) -> u64;
    fn tx_chain_id_check(&self) -> bool;
    fn tx_gas_limit_cap(&self) -> u64;
    fn spec(&self) -> Self::Spec;
    fn max_blobs_per_tx(&self) -> Option<u64>;
    fn max_code_size(&self) -> usize;
    fn max_initcode_size(&self) -> usize;
    fn is_eip3607_disabled(&self) -> bool;
    fn is_eip3541_disabled(&self) -> bool;
    fn is_balance_check_disabled(&self) -> bool;
    fn is_block_gas_limit_disabled(&self) -> bool;
    fn is_nonce_check_disabled(&self) -> bool;
    fn is_base_fee_check_disabled(&self) -> bool;
    fn is_priority_fee_check_disabled(&self) -> bool;

}

Configuration for the EVM.

Required Associated Types

type Spec: Into<SpecId> + Clone

Specification id type, it requires to be convertible to SpecId so it can be used by default in mainnet.

Required Methods

fn chain_id(&self) -> u64

Returns the chain ID of the EVM that is compared with the transaction’s chain ID.

fn tx_chain_id_check(&self) -> bool

Returns whether the transaction’s chain ID check is enabled.

fn tx_gas_limit_cap(&self) -> u64

Returns the gas limit cap for the transaction.

Cap is introduced in EIP-7825: Transaction Gas Limit Cap with initial cap of 30M gas.

Value before EIP-7825 is u64::MAX.

fn spec(&self) -> Self::Spec

Specification id

fn max_blobs_per_tx(&self) -> Option<u64>

Returns the maximum number of blobs allowed per transaction. If it is None, check for max count will be skipped.

fn max_code_size(&self) -> usize

Returns the maximum code size for the given spec id.

fn max_initcode_size(&self) -> usize

Returns the max initcode size for the given spec id.

fn is_eip3607_disabled(&self) -> bool

Returns whether the EIP-3607 (account clearing) is disabled.

fn is_eip3541_disabled(&self) -> bool

Returns whether the EIP-3541 (disallowing new contracts with 0xEF prefix) is disabled.

fn is_balance_check_disabled(&self) -> bool

Returns whether the balance check is disabled.

fn is_block_gas_limit_disabled(&self) -> bool

Returns whether the block gas limit check is disabled.

fn is_nonce_check_disabled(&self) -> bool

Returns whether the nonce check is disabled.

fn is_base_fee_check_disabled(&self) -> bool

Returns whether the base fee check is disabled.

fn is_priority_fee_check_disabled(&self) -> bool

Returns whether the priority fee check is disabled.