cast::revm::context_interface

Module local

Local context trait LocalContextTr and related types.

Structs

FrameStack

Non-empty, item-pooling Vec.

FrameToken

Used to guarantee that a frame is initialized before use.

OutFrame

A potentially initialized frame. Used when initializing a new frame in the main loop.

Traits

LocalContextTr

Local context used for caching initcode from Initcode transactions.