cast::revm::handler::api

Trait ExecuteEvm

pub trait ExecuteEvm {
    type Output;
    type Tx: Transaction;
    type Block: Block;

    // Required methods
    fn set_tx(&mut self, tx: Self::Tx);
    fn set_block(&mut self, block: Self::Block);
    fn replay(&mut self) -> Self::Output;

    // Provided method
    fn transact(&mut self, tx: Self::Tx) -> Self::Output { ... }
}

Execute EVM transactions. Main trait for transaction execution.

Required Associated Types

type Output

Output of transaction execution.

type Tx: Transaction

Transaction type.

type Block: Block

Block type.

Required Methods

fn set_tx(&mut self, tx: Self::Tx)

Set the transaction.

fn set_block(&mut self, block: Self::Block)

Set the block.

fn replay(&mut self) -> Self::Output

Transact the transaction that is set in the context.

Provided Methods

fn transact(&mut self, tx: Self::Tx) -> Self::Output

Transact the given transaction.

Internally sets transaction in context and use replay to execute the transaction.