cast::revm::handler::system_call

Trait SystemCallCommitEvm

pub trait SystemCallCommitEvm: SystemCallEvm + ExecuteCommitEvm {
    // Required method
    fn transact_system_call_commit(
        &mut self,
        system_contract_address: Address,
        data: Bytes,
    ) -> Self::CommitOutput;
}

Extension of the SystemCallEvm trait that adds a method that commits the state after execution.

Required Methods

fn transact_system_call_commit( &mut self, system_contract_address: Address, data: Bytes, ) -> Self::CommitOutput

Transact the system call and commit to the state.