cast::revm::handler::system_call

Trait SystemCallEvm

pub trait SystemCallEvm: ExecuteEvm {
    // Required method
    fn transact_system_call(
        &mut self,
        system_contract_address: Address,
        data: Bytes,
    ) -> Self::Output;
}

API for executing the system calls. System calls dont deduct the caller or reward the beneficiary. They are used before and after block execution to insert or obtain blockchain state.

It act similar to transact function and sets default Tx with data and system contract as a target.

Required Methods

fn transact_system_call( &mut self, system_contract_address: Address, data: Bytes, ) -> Self::Output

System call is a special transaction call that is used to call a system contract.

Transaction fields are reset and set in SystemCallTx and data and target are set to given values.

Block values are taken into account and will determent how system call will be executed.