cast::revm::handler::system_call

Trait SystemCallTx

pub trait SystemCallTx: Sized {
    // Required method
    fn new_system_tx_with_caller(
        caller: Address,
        system_contract_address: Address,
        data: Bytes,
    ) -> Self;

    // Provided method
    fn new_system_tx(system_contract_address: Address, data: Bytes) -> Self { ... }
}

Creates the system transaction with default values and set data and tx call target to system contract address that is going to be called.

The caller is set to be SYSTEM_ADDRESS.

It is used inside SystemCallEvm and SystemCallCommitEvm traits to prepare EVM for system call execution.

Required Methods

fn new_system_tx_with_caller( caller: Address, system_contract_address: Address, data: Bytes, ) -> Self

Creates a new system transaction with a custom caller address.

Provided Methods

fn new_system_tx(system_contract_address: Address, data: Bytes) -> Self

Creates new transaction for system call.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.