cast::revm::handler

Trait FrameTr

pub trait FrameTr {
    type FrameResult: Into<FrameResult>;
    type FrameInit: Into<FrameInit>;
}

Trait for defining a frame type used in EVM execution.

Required Associated Types

type FrameResult: Into<FrameResult>

The result type returned when a frame completes execution.

type FrameInit: Into<FrameInit>

The initialization type used to create a new frame.