cast::revm::inspector

Trait Inspector

pub trait Inspector<CTX, INTR = EthInterpreter>
where
    INTR: InterpreterTypes,
{
    // Provided methods
    fn initialize_interp(
        &mut self,
        interp: &mut Interpreter<INTR>,
        context: &mut CTX,
    ) { ... }
    fn step(&mut self, interp: &mut Interpreter<INTR>, context: &mut CTX) { ... }
    fn step_end(&mut self, interp: &mut Interpreter<INTR>, context: &mut CTX) { ... }
    fn log(
        &mut self,
        interp: &mut Interpreter<INTR>,
        context: &mut CTX,
        log: Log,
    ) { ... }
    fn call(
        &mut self,
        context: &mut CTX,
        inputs: &mut CallInputs,
    ) -> Option<CallOutcome> { ... }
    fn call_end(
        &mut self,
        context: &mut CTX,
        inputs: &CallInputs,
        outcome: &mut CallOutcome,
    ) { ... }
    fn create(
        &mut self,
        context: &mut CTX,
        inputs: &mut CreateInputs,
    ) -> Option<CreateOutcome> { ... }
    fn create_end(
        &mut self,
        context: &mut CTX,
        inputs: &CreateInputs,
        outcome: &mut CreateOutcome,
    ) { ... }
    fn selfdestruct(
        &mut self,
        contract: Address,
        target: Address,
        value: Uint<256, 4>,
    ) { ... }
}

EVM hooks into execution.

This trait is used to enabled tracing of the EVM execution.

Object that is implemented this trait is used in InspectorHandler to trace the EVM execution. And API that allow calling the inspector can be found in crate::InspectEvm and crate::InspectCommitEvm.

Provided Methods

fn initialize_interp( &mut self, interp: &mut Interpreter<INTR>, context: &mut CTX, )

Called before the interpreter is initialized.

If interp.bytecode.set_action is set the execution of the interpreter is skipped.

fn step(&mut self, interp: &mut Interpreter<INTR>, context: &mut CTX)

Called on each step of the interpreter.

Information about the current execution, including the memory, stack and more is available on interp (see Interpreter).

Example

To get the current opcode, use interp.bytecode.opcode().

fn step_end(&mut self, interp: &mut Interpreter<INTR>, context: &mut CTX)

Called after step when the instruction has been executed.

Setting interp.bytecode.set_action will result in stopping the execution of the interpreter.

fn log(&mut self, interp: &mut Interpreter<INTR>, context: &mut CTX, log: Log)

Called when a log is emitted.

fn call( &mut self, context: &mut CTX, inputs: &mut CallInputs, ) -> Option<CallOutcome>

Called whenever a call to a contract is about to start.

Returning CallOutcome will override the result of the call.

fn call_end( &mut self, context: &mut CTX, inputs: &CallInputs, outcome: &mut CallOutcome, )

Called when a call to a contract has concluded.

The returned CallOutcome is used as the result of the call.

This allows the inspector to modify the given result before returning it.

fn create( &mut self, context: &mut CTX, inputs: &mut CreateInputs, ) -> Option<CreateOutcome>

Called when a contract is about to be created.

If this returns Some then the CreateOutcome is used to override the result of the creation.

If this returns None then the creation proceeds as normal.

fn create_end( &mut self, context: &mut CTX, inputs: &CreateInputs, outcome: &mut CreateOutcome, )

Called when a contract has been created.

Modifying the outcome will alter the result of the create operation.

fn selfdestruct( &mut self, contract: Address, target: Address, value: Uint<256, 4>, )

Called when a contract has been self-destructed with funds transferred to target.