cast::revm::interpreter

Module interpreter

Modules

ext_bytecode

Structs

EthInterpreter

Default types for Ethereum interpreter.

ExtBytecode

InputsImpl

Inputs for the interpreter that are used for execution of the call.

Interpreter

Main interpreter structure that contains all components defines in InterpreterTypes.s

InterpreterResult

The result of an interpreter operation.

LoopControlImpl

ReturnDataImpl

RuntimeFlags

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack with STACK_LIMIT capacity of words.

SubRoutineImpl

Function Stack

SubRoutineReturnFrame

Function(Sub Routine) return frame in eof

Constants

STACK_LIMIT

EVM interpreter stack limit.

Functions

num_words

Returns number of words what would fit to provided number of bytes, i.e. it rounds up the number bytes to number of words.