Struct cast::revm::interpreter::interpreter_action::CreateOutcome

pub struct CreateOutcome {
    pub result: InterpreterResult,
    pub address: Option<Address>,
}

Represents the outcome of a create operation in an interpreter.

This struct holds the result of the operation along with an optional address. It provides methods to determine the next action based on the result of the operation.

Fields
result: InterpreterResultaddress: Option<Address>
Implementations

impl CreateOutcome

pub fn new(result: InterpreterResult, address: Option<Address>) -> CreateOutcome

Constructs a new CreateOutcome.

Arguments

result - An InterpreterResult representing the result of the interpreter operation.

address - An optional Address associated with the create operation.

Returns

A new CreateOutcome instance.

pub fn instruction_result(&self) -> &InstructionResult

Retrieves a reference to the InstructionResult from the InterpreterResult.

This method provides access to the InstructionResult which represents the outcome of the instruction execution. It encapsulates the result information such as whether the instruction was executed successfully, resulted in a revert, or encountered a fatal error.

Returns

A reference to the InstructionResult.

pub fn output(&self) -> &Bytes

Retrieves a reference to the output bytes from the InterpreterResult.

This method returns the output of the interpreted operation. The output is typically used when the operation successfully completes and returns data.

Returns

A reference to the output Bytes.

pub fn gas(&self) -> &Gas

Retrieves a reference to the Gas details from the InterpreterResult.

This method provides access to the gas details of the operation, which includes information about gas used, remaining, and refunded. It is essential for understanding the gas consumption of the operation.

Returns

A reference to the Gas details.