cast::revm::interpreter

Module interpreter_types

Type traits and definitions for interpreter customization.

Traits

Immediates

Helper function to read immediates data from the bytecode

InputsTr

Trait for fetching inputs of the call.

Interp

Trait for interpreter execution.

InterpreterTypes

Trait defining the component types used by an interpreter implementation.

Jumps

Trait for Interpreter to be able to jump

LegacyBytecode

Trait needed for legacy bytecode.

LoopControl

Trait controls execution of the loop.

MemoryTr

Trait for Interpreter memory operations.

ReturnData

Returns return data.

RuntimeFlag

Runtime flags that control interpreter execution behavior.

StackTr

Functions needed for Interpreter Stack operations.