Crate cast::revm::interpreter::primitives

revm-primitives

EVM primitive types.

Modules

alloy_primitives

alloy-primitives

bitvec

db

env

precompile

result

specification

state

utilities

Macros

address

Converts a sequence of string literals containing hex-encoded data into a new Address at compile time.

b256

Converts a sequence of string literals containing hex-encoded data into a new B256 at compile time.

bytes

Converts a sequence of string literals containing hex-encoded data into a new Bytes at compile time.

fixed_bytes

Converts a sequence of string literals containing hex-encoded data into a new FixedBytes at compile time.

hex

Macro for converting sequence of string literals containing hex-encoded data into an array of bytes.

spec_to_generic

Structs

Account

AccountInfo

AccountInfo account information.

AccountStatus

Address

An Ethereum address, 20 bytes in length.

BedrockSpec

BerlinSpec

BlobExcessGasAndPrice

Structure holding block blob excess gas and it calculates blob fee.

BlockEnv

The block environment.

Bytecode

Bytes

Wrapper type around [bytes::Bytes] to support “0x” prefixed hex strings.

ByzantiumSpec

CancunSpec

CanyonSpec

CfgEnv

EVM configuration.

CfgEnvWithHandlerCfg

Configuration environment with the chain spec id.

EcotoneSpec

Env

EVM environment configuration.

EnvWithHandlerCfg

Evm environment with the chain spec id.

FixedBytes

A byte array of fixed length ([u8; N]).

FrontierSpec

HandlerCfg

Handler configuration fields. It is used to configure the handler. It contains specification id and the Optimism related field if optimism feature is enabled.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

HomesteadSpec

IstanbulSpec

JumpMap

A map of valid jump destinations.

LatestSpec

Log

A log consists of an address, and some log data.

LogData

An Ethereum event log object.

LondonSpec

MergeSpec

OptimismFields

Additional TxEnv fields for optimism.

PetersburgSpec

RegolithSpec

ResultAndState

ShanghaiSpec

SpuriousDragonSpec

StorageSlot

This type keeps track of the current value of a storage slot.

TangerineSpec

TxEnv

The transaction environment.

Enums

AnalysisKind

What bytecode analysis to perform.

BytecodeState

State of the Bytecode analysis.

CreateScheme

Create scheme.

EVMError

Main EVM error.

ExecutionResult

Result of a transaction execution.

HaltReason

Indicates that the EVM has experienced an exceptional halt. This causes execution to immediately end with all gas being consumed.

InvalidHeader

Errors related to misconfiguration of a crate::env::BlockEnv.

InvalidTransaction

Transaction validation error.

OutOfGasError

Output

Output of a transaction execution.

Precompile

Precompile and its handlers.

PrecompileError

SpecId

Specification IDs and their activation block.

SuccessReason

Reason a transaction successfully completed.

TransactTo

Transaction destination.

Constants

BLOB_GASPRICE_UPDATE_FRACTION

Controls the maximum rate of change for blob gas price.

BLOCK_HASH_HISTORY

Number of block hashes that EVM can access in the past

GAS_PER_BLOB

Gas consumption of a single data blob (== blob byte size).

KECCAK_EMPTY

The Keccak-256 hash of the empty string "".

MAX_BLOB_GAS_PER_BLOCK

Maximum consumable blob gas for data blobs per block.

MAX_BLOB_NUMBER_PER_BLOCK

Max number of blobs per block

MAX_CODE_SIZE

EIP-170: Contract code size limit By default limit is 0x6000 (~25kb)

MAX_INITCODE_SIZE

EIP-3860: Limit and meter initcode

MIN_BLOB_GASPRICE

Minimum gas price for data blobs.

PRECOMPILE3

Precompile 3 is special in few places

TARGET_BLOB_GAS_PER_BLOCK

Target consumable blob gas for data blobs per block (for 1559-like pricing).

TARGET_BLOB_NUMBER_PER_BLOCK

Target number of the blob per block.

VERSIONED_HASH_VERSION_KZG

First version of the blob.

Traits

Spec

StatefulPrecompile

Stateful precompile trait. It is used to create a arc precompile Precompile::Stateful.

StatefulPrecompileMut

Mutable stateful precompile trait. It is used to create a boxed precompile in Precompile::StatefulMut.

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

fake_exponential

Approximates factor * e ** (numerator / denominator) using Taylor expansion.

keccak256

Simple interface to the Keccak-256 hash function.

Type Aliases

B256

32-byte fixed byte-array type.

EVMResult

Result of EVM execution.

EVMResultGeneric

Generic result of EVM execution. Used to represent error and generic output.

EnvPrecompileFn

I256

256-bit signed integer type, consisting of 4, 64-bit limbs.

PrecompileResult

A precompile operation result.

StandardPrecompileFn

State

EVM State is a mapping from addresses to accounts.

StatefulPrecompileArc

Arc over stateful precompile.

StatefulPrecompileBox

Box over mutable stateful precompile

Storage

An account’s Storage is a mapping from 256-bit integer keys to StorageSlots.

TransientStorage

Structure used for EIP-1153 transient storage.

U256

Uint for 256 bits.