Module cast::revm::interpreter::primitives::alloy_primitives::aliases

Type aliases for common primitive types.

Type Aliases

B8

1-byte fixed byte-array type.

B16

2-byte fixed byte-array type.

B32

4-byte fixed byte-array type.

B64

8-byte fixed byte-array type.

B96

12-byte fixed byte-array type.

B128

16-byte fixed byte-array type.

B192

24-byte fixed byte-array type.

B224

28-byte fixed byte-array type.

B256

32-byte fixed byte-array type.

B512

64-byte fixed byte-array type.

B1024

128-byte fixed byte-array type.

B2048

256-byte fixed byte-array type.

BlockHash

A block hash.

BlockNumber

A block number.

BlockTimestamp

A block timestamp.

ChainId

Chain identifier type (introduced in EIP-155).

I0

The 0-bit signed integer type, capable of representing 0.

I1

The 1-bit signed integer type, capable of representing 0 and -1.

I8

8-bit signed integer type, consisting of 1, 64-bit limbs.

I16

16-bit signed integer type, consisting of 1, 64-bit limbs.

I32

32-bit signed integer type, consisting of 1, 64-bit limbs.

I64

64-bit signed integer type, consisting of 1, 64-bit limbs.

I128

128-bit signed integer type, consisting of 2, 64-bit limbs.

I160

160-bit signed integer type, consisting of 3, 64-bit limbs.

I192

192-bit signed integer type, consisting of 3, 64-bit limbs.

I256

256-bit signed integer type, consisting of 4, 64-bit limbs.

I512

512-bit signed integer type, consisting of 8, 64-bit limbs.

Selector

Solidity contract functions are addressed using the first four bytes of the Keccak-256 hash of their signature.

StorageKey

An account storage key.

StorageValue

An account storage value.

TxHash

A transaction hash is a keccak hash of an RLP encoded signed transaction.

TxIndex

The index of transaction in a block.

TxNonce

The nonce of a transaction.

TxNumber

The sequence number of all existing transactions.

U0

Uint for 0 bits. Always zero. Similar to ().

U1

Uint for 1 bit. Similar to bool.

U8

Uint for 8 bits. Similar to u8.

U16

Uint for 16 bits. Similar to u16.

U32

Uint for 32 bits. Similar to u32.

U64

Uint for 64 bits. Similar to u64.

U128

Uint for 128 bits. Similar to u128.

U160

Uint for 160 bits.

U192

Uint for 192 bits.

U256

Uint for 256 bits.

U320

Uint for 320 bits.

U384

Uint for 384 bits.

U448

Uint for 448 bits.

U512

Uint for 512 bits.

U1024

Uint for 1024 bits.

U2048

Uint for 2048 bits.

U4096

Uint for 4096 bits.