cast::revm::interpreter::primitives

Module env

Modules

handler_cfg

Structs

BlobExcessGasAndPrice

Structure holding block blob excess gas and it calculates blob fee.

BlockEnv

The block environment.

CfgEnv

EVM configuration.

CfgEnvWithHandlerCfg

Configuration environment with the chain spec id.

Env

EVM environment configuration.

EnvWithHandlerCfg

Evm environment with the chain spec id.

HandlerCfg

Handler configuration fields. It is used to configure the handler. It contains specification id and the Optimism related field if optimism feature is enabled.

OptimismFieldsoptimism

Additional TxEnv fields for optimism.

TxEnv

The transaction environment.

Enums

AnalysisKind

What bytecode analysis to perform.

CreateScheme

Create scheme.

TxType

Type Aliases

TransactTo

Transaction destination