Module cast::revm::interpreter::primitives::precompile

Structs

PrecompileOutput

Precompile execution output

Enums

Precompile

Precompile and its handlers.

PrecompileError

PrecompileErrors

Precompile errors.

Traits

StatefulPrecompile

Stateful precompile trait. It is used to create a arc precompile Precompile::Stateful.

StatefulPrecompileMut

Mutable stateful precompile trait. It is used to create a boxed precompile in Precompile::StatefulMut.

Type Aliases

EnvPrecompileFn

PrecompileResult

A precompile operation result.

StandardPrecompileFn

StatefulPrecompileArc

Arc over stateful precompile.

StatefulPrecompileBox

Box over mutable stateful precompile