Trait cast::revm::interpreter::Host

pub trait Host {

    // Required methods
    fn env(&self) -> &Env;
    fn env_mut(&mut self) -> &mut Env;
    fn load_account(&mut self, address: Address) -> Option<(bool, bool)>;
    fn block_hash(&mut self, number: Uint<256, 4>) -> Option<FixedBytes<32>>;
    fn balance(&mut self, address: Address) -> Option<(Uint<256, 4>, bool)>;
    fn code(&mut self, address: Address) -> Option<(Bytecode, bool)>;
    fn code_hash(&mut self, address: Address) -> Option<(FixedBytes<32>, bool)>;
    fn sload(
        &mut self,
        address: Address,
        index: Uint<256, 4>
    ) -> Option<(Uint<256, 4>, bool)>;
    fn sstore(
        &mut self,
        address: Address,
        index: Uint<256, 4>,
        value: Uint<256, 4>
    ) -> Option<SStoreResult>;
    fn tload(&mut self, address: Address, index: Uint<256, 4>) -> Uint<256, 4>;
    fn tstore(
        &mut self,
        address: Address,
        index: Uint<256, 4>,
        value: Uint<256, 4>
    );
    fn log(&mut self, log: Log);
    fn selfdestruct(
        &mut self,
        address: Address,
        target: Address
    ) -> Option<SelfDestructResult>;

}

EVM context host.

Required Methods

fn env(&self) -> &Env

Returns a reference to the environment.

fn env_mut(&mut self) -> &mut Env

Returns a mutable reference to the environment.

fn load_account(&mut self, address: Address) -> Option<(bool, bool)>

Load an account.

Returns (is_cold, is_new_account)

fn block_hash(&mut self, number: Uint<256, 4>) -> Option<FixedBytes<32>>

Get the block hash of the given block number.

fn balance(&mut self, address: Address) -> Option<(Uint<256, 4>, bool)>

Get balance of address and if the account is cold.

fn code(&mut self, address: Address) -> Option<(Bytecode, bool)>

Get code of address and if the account is cold.

fn code_hash(&mut self, address: Address) -> Option<(FixedBytes<32>, bool)>

Get code hash of address and if the account is cold.

fn sload( &mut self, address: Address, index: Uint<256, 4> ) -> Option<(Uint<256, 4>, bool)>

Get storage value of address at index and if the account is cold.

fn sstore( &mut self, address: Address, index: Uint<256, 4>, value: Uint<256, 4> ) -> Option<SStoreResult>

Set storage value of account address at index.

Returns (original, present, new, is_cold).

fn tload(&mut self, address: Address, index: Uint<256, 4>) -> Uint<256, 4>

Get the transient storage value of address at index.

fn tstore(&mut self, address: Address, index: Uint<256, 4>, value: Uint<256, 4>)

Set the transient storage value of address at index.

fn log(&mut self, log: Log)

Emit a log owned by address with given LogData.

fn selfdestruct( &mut self, address: Address, target: Address ) -> Option<SelfDestructResult>

Mark address to be deleted, with funds transferred to target.