Crate cast::revm::precompile

revm-precompile

Implementations of EVM precompiled contracts.

Modules

utilities

Structs

Address

An Ethereum address, 20 bytes in length.

Bytes

Wrapper type around [bytes::Bytes] to support “0x” prefixed hex strings.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

Log

A log consists of an address, and some log data.

PrecompileOutput

PrecompileWithAddress

Precompiles

Enums

Error

Precompile

Precompile and its handlers.

PrecompileError

PrecompileSpecId

Traits

StatefulPrecompile

Stateful precompile trait. It is used to create a arc precompile Precompile::Stateful.

StatefulPrecompileMut

Mutable stateful precompile trait. It is used to create a boxed precompile in Precompile::StatefulMut.

Functions

calc_linear_cost_u32

Type Aliases

B256

32-byte fixed byte-array type.

EnvPrecompileFn

PrecompileResult

A precompile operation result.

StandardPrecompileFn

StatefulPrecompileArc

Arc over stateful precompile.

StatefulPrecompileBox

Box over mutable stateful precompile