cast::revm::precompile

Module secp256r1

RIP-7212 secp256r1 Precompile

This module implements the RIP-7212 precompile for secp256r1 curve support.

The main purpose of this precompile is to verify ECDSA signatures that use the secp256r1, or P256 elliptic curve. The P256VERIFY const represents the implementation of this precompile, with the address that it is currently deployed at.

Constants

P256VERIFY

RIP-7212 secp256r1 precompile.

P256VERIFY_ADDRESS

Address of secp256r1 precompile.

P256VERIFY_BASE_GAS_FEE

Base gas fee for secp256r1 p256verify operation.

P256VERIFY_BASE_GAS_FEE_OSAKA

Base gas fee for secp256r1 p256verify operation post Osaka.

P256VERIFY_OSAKA

RIP-7212 secp256r1 precompile.

Functions

p256_verify

secp256r1 precompile logic. It takes the input bytes sent to the precompile and the gas limit. The output represents the result of verifying the secp256r1 signature of the input.

p256_verify_osaka

secp256r1 precompile logic with Osaka gas cost. It takes the input bytes sent to the precompile and the gas limit. The output represents the result of verifying the secp256r1 signature of the input.

precompiles

Returns the secp256r1 precompile with its address.

verify_impl

Returns Some(()) if the signature included in the input byte slice is valid, None otherwise.