Struct cast::revm::precompile::Bytes

#[repr(transparent)]
pub struct Bytes(pub Bytes);

Wrapper type around [bytes::Bytes] to support โ0xโ prefixed hex strings.

Tuple Fields
0: Bytes
Implementations

impl Bytes

pub const fn new() -> Bytes

Creates a new empty Bytes.

This will not allocate and the returned Bytes handle will be empty.

Examples

use alloy_primitives::Bytes;

let b = Bytes::new();
assert_eq!(&b[..], b"");

pub const fn from_static(bytes: &'static [u8]) -> Bytes

Creates a new Bytes from a static slice.

The returned Bytes will point directly to the static slice. There is no allocating or copying.

Examples

use alloy_primitives::Bytes;

let b = Bytes::from_static(b"hello");
assert_eq!(&b[..], b"hello");

pub fn copy_from_slice(data: &[u8]) -> Bytes

Creates a new Bytes instance from a slice by copying it.

pub fn slice(&self, range: impl RangeBounds<usize>) -> Bytes

Returns a slice of self for the provided range.

pub fn slice_ref(&self, subset: &[u8]) -> Bytes

Returns a slice of self that is equivalent to the given subset.

pub fn split_off(&mut self, at: usize) -> Bytes

Splits the bytes into two at the given index.

pub fn split_to(&mut self, at: usize) -> Bytes

Splits the bytes into two at the given index.