cast::revm::primitives::alloy_primitives::ruint

Trait UintTryFrom

pub trait UintTryFrom<T>: Sized { }

⚠️ Workaround for Rust issue #50133. Use TryFrom instead.

We cannot implement TryFrom<Uint> for Uint directly, but we can create a new identical trait and implement it there. We can even give this trait a blanket implementation inheriting all TryFrom<_> implementations.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.