cast::revm::primitives::alloy_primitives

Struct Function

#[repr(transparent)]
pub struct Function(pub FixedBytes<24>);

An Ethereum ABI function pointer, 24 bytes in length.

An address (20 bytes), followed by a function selector (4 bytes). Encoded identical to bytes24.

Tuple Fields
0: FixedBytes<24>
Implementations

impl Function

pub const ZERO: Function

Array of Zero bytes.

pub const fn new(bytes: [u8; 24]) -> Function

Wraps the given byte array in this type.

pub const fn with_last_byte(x: u8) -> Function

Creates a new byte array with the last byte set to x.

pub const fn repeat_byte(byte: u8) -> Function

Creates a new byte array where all bytes are set to byte.

pub const fn len_bytes() -> usize

Returns the size of this array in bytes.

pub fn random() -> Function

Creates a new fixed byte array with the default cryptographic random number generator.

This is rand::thread_rng if the “rand” and “std” features are enabled, otherwise it uses getrandom::getrandom. Both are cryptographically secure.

pub fn try_random() -> Result<Function, Error>

Tries to create a new fixed byte array with the default cryptographic random number generator.

See random for more details.

pub fn randomize(&mut self)

Fills this fixed byte array with the default cryptographic random number generator.

See random for more details.

pub fn try_randomize(&mut self) -> Result<(), Error>

Tries to fill this fixed byte array with the default cryptographic random number generator.

See random for more details.

pub fn random_with<R>(rng: &mut R) -> Function
where R: RngCore + ?Sized,

Creates a new fixed byte array with the given random number generator.

pub fn try_random_with<R>( rng: &mut R, ) -> Result<Function, <R as TryRngCore>::Error>
where R: TryRngCore + ?Sized,

Tries to create a new fixed byte array with the given random number generator.

pub fn randomize_with<R>(&mut self, rng: &mut R)
where R: RngCore + ?Sized,

Fills this fixed byte array with the given random number generator.

pub fn try_randomize_with<R>( &mut self, rng: &mut R, ) -> Result<(), <R as TryRngCore>::Error>
where R: TryRngCore + ?Sized,

Tries to fill this fixed byte array with the given random number generator.

pub fn from_slice(src: &[u8]) -> Function

Create a new byte array from the given slice src.

For a fallible version, use the TryFrom<&[u8]> implementation.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn left_padding_from(value: &[u8]) -> Function

Create a new byte array from the given slice src, left-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub fn right_padding_from(value: &[u8]) -> Function

Create a new byte array from the given slice src, right-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub const fn into_array(self) -> [u8; 24]

Returns the inner bytes array.

pub fn covers(&self, b: &Function) -> bool

Returns true if all bits set in b are also set in self.

pub const fn const_eq(&self, other: &Function) -> bool

Compile-time equality. NOT constant-time equality.

pub const fn bit_and(self, rhs: Function) -> Function

Computes the bitwise AND of two FixedBytes.

pub const fn bit_or(self, rhs: Function) -> Function

Computes the bitwise OR of two FixedBytes.

pub const fn bit_xor(self, rhs: Function) -> Function

Computes the bitwise XOR of two FixedBytes.

impl Function

pub fn from_word(word: FixedBytes<32>) -> Function

Creates an Ethereum function from an EVM word’s lower 24 bytes (word[..24]).

Note that this is different from Address::from_word, which uses the upper 20 bytes.

pub fn into_word(&self) -> FixedBytes<32>

Right-pads the function to 32 bytes (EVM word size).

Note that this is different from Address::into_word, which left-pads the address.

pub fn from_address_and_selector<A, S>(address: A, selector: S) -> Function
where A: Borrow<[u8; 20]>, S: Borrow<[u8; 4]>,

Creates an Ethereum function from an address and selector.

pub fn as_address_and_selector(&self) -> (&Address, &FixedBytes<4>)

Returns references to the address and selector of the function.

pub fn to_address_and_selector(&self) -> (Address, FixedBytes<4>)

Returns the address and selector of the function.