cast::revm::primitives::map::indexmap::map

Struct Slice

#[repr(transparent)]
pub struct Slice<K, V> {
    pub(crate) entries: [Bucket<K, V>],
}

A dynamically-sized slice of key-value pairs in an IndexMap.

This supports indexed operations much like a [(K, V)] slice, but not any hashed operations on the map keys.

Unlike IndexMap, Slice does consider the order for PartialEq and Eq, and it also implements PartialOrd, Ord, and Hash.

Fields
entries: [Bucket<K, V>]
Implementations

impl<K, V> Slice<K, V>

pub const fn new<'a>() -> &'a Slice<K, V>

Returns an empty slice.

pub fn new_mut<'a>() -> &'a mut Slice<K, V>

Returns an empty mutable slice.

pub const fn len(&self) -> usize

Return the number of key-value pairs in the map slice.

pub const fn is_empty(&self) -> bool

Returns true if the map slice contains no elements.

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Get a key-value pair by index.

Valid indices are 0 <= index < self.len().

pub fn get_index_mut(&mut self, index: usize) -> Option<(&K, &mut V)>

Get a key-value pair by index, with mutable access to the value.

Valid indices are 0 <= index < self.len().

pub fn get_range<R>(&self, range: R) -> Option<&Slice<K, V>>
where R: RangeBounds<usize>,

Returns a slice of key-value pairs in the given range of indices.

Valid indices are 0 <= index < self.len().

pub fn get_range_mut<R>(&mut self, range: R) -> Option<&mut Slice<K, V>>
where R: RangeBounds<usize>,

Returns a mutable slice of key-value pairs in the given range of indices.

Valid indices are 0 <= index < self.len().

pub fn first(&self) -> Option<(&K, &V)>

Get the first key-value pair.

pub fn first_mut(&mut self) -> Option<(&K, &mut V)>

Get the first key-value pair, with mutable access to the value.

pub fn last(&self) -> Option<(&K, &V)>

Get the last key-value pair.

pub fn last_mut(&mut self) -> Option<(&K, &mut V)>

Get the last key-value pair, with mutable access to the value.

pub fn split_at(&self, index: usize) -> (&Slice<K, V>, &Slice<K, V>)

Divides one slice into two at an index.

Panics if index > len.

pub fn split_at_mut( &mut self, index: usize, ) -> (&mut Slice<K, V>, &mut Slice<K, V>)

Divides one mutable slice into two at an index.

Panics if index > len.

pub fn split_first(&self) -> Option<((&K, &V), &Slice<K, V>)>

Returns the first key-value pair and the rest of the slice, or None if it is empty.

pub fn split_first_mut(&mut self) -> Option<((&K, &mut V), &mut Slice<K, V>)>

Returns the first key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

pub fn split_last(&self) -> Option<((&K, &V), &Slice<K, V>)>

Returns the last key-value pair and the rest of the slice, or None if it is empty.

pub fn split_last_mut(&mut self) -> Option<((&K, &mut V), &mut Slice<K, V>)>

Returns the last key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

pub fn iter(&self) -> Iter<'_, K, V>

Return an iterator over the key-value pairs of the map slice.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Return an iterator over the key-value pairs of the map slice.

pub fn keys(&self) -> Keys<'_, K, V>

Return an iterator over the keys of the map slice.

pub fn into_keys(self: Box<Slice<K, V>>) -> IntoKeys<K, V>

Return an owning iterator over the keys of the map slice.

pub fn values(&self) -> Values<'_, K, V>

Return an iterator over the values of the map slice.

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>

Return an iterator over mutable references to the the values of the map slice.

pub fn into_values(self: Box<Slice<K, V>>) -> IntoValues<K, V>

Return an owning iterator over the values of the map slice.

pub fn binary_search_keys(&self, x: &K) -> Result<usize, usize>
where K: Ord,

Search over a sorted map for a key.

Returns the position where that key is present, or the position where it can be inserted to maintain the sort. See slice::binary_search for more details.

Computes in O(log(n)) time, which is notably less scalable than looking the key up in the map this is a slice from using IndexMap::get_index_of, but this can also position missing keys.

pub fn binary_search_by<'a, F>(&'a self, f: F) -> Result<usize, usize>
where F: FnMut(&'a K, &'a V) -> Ordering,

Search over a sorted map with a comparator function.

Returns the position where that value is present, or the position where it can be inserted to maintain the sort. See slice::binary_search_by for more details.

Computes in O(log(n)) time.

pub fn binary_search_by_key<'a, B, F>( &'a self, b: &B, f: F, ) -> Result<usize, usize>
where F: FnMut(&'a K, &'a V) -> B, B: Ord,

Search over a sorted map with an extraction function.

Returns the position where that value is present, or the position where it can be inserted to maintain the sort. See slice::binary_search_by_key for more details.

Computes in O(log(n)) time.

pub fn is_sorted(&self) -> bool
where K: PartialOrd,

Checks if the keys of this slice are sorted.

pub fn is_sorted_by<'a, F>(&'a self, cmp: F) -> bool
where F: FnMut(&'a K, &'a V, &'a K, &'a V) -> bool,

Checks if this slice is sorted using the given comparator function.

pub fn is_sorted_by_key<'a, F, T>(&'a self, sort_key: F) -> bool
where F: FnMut(&'a K, &'a V) -> T, T: PartialOrd,

Checks if this slice is sorted using the given sort-key function.

pub fn partition_point<P>(&self, pred: P) -> usize
where P: FnMut(&K, &V) -> bool,

Returns the index of the partition point of a sorted map according to the given predicate (the index of the first element of the second partition).

See slice::partition_point for more details.

Computes in O(log(n)) time.

pub fn get_disjoint_mut<const N: usize>( &mut self, indices: [usize; N], ) -> Result<[(&K, &mut V); N], GetDisjointMutError>

Get an array of N key-value pairs by N indices

Valid indices are 0 <= index < self.len() and each index needs to be unique.