Struct cast::revm::JournaledState

pub struct JournaledState {
    pub state: HashMap<Address, Account>,
    pub transient_storage: HashMap<(Address, Uint<256, 4>), Uint<256, 4>>,
    pub logs: Vec<Log>,
    pub depth: usize,
    pub journal: Vec<Vec<JournalEntry>>,
    pub spec: SpecId,
    pub warm_preloaded_addresses: HashSet<Address>,
}

JournalState is internal EVM state that is used to contain state and track changes to that state. It contains journal of changes that happened to state so that they can be reverted.

Fields
state: HashMap<Address, Account>

Current state.

transient_storage: HashMap<(Address, Uint<256, 4>), Uint<256, 4>>

EIP-1153 transient storage that is discarded after every transactions

logs: Vec<Log>

logs

depth: usize

how deep are we in call stack.

journal: Vec<Vec<JournalEntry>>

journal with changes that happened between calls.

spec: SpecId

Ethereum before EIP-161 differently defined empty and not-existing account Spec is needed for two things SpuriousDragon’s EIP-161 State clear, and for Cancun’s EIP-6780: SELFDESTRUCT in same transaction

warm_preloaded_addresses: HashSet<Address>

Warm loaded addresses are used to check if loaded address should be considered cold or warm loaded when the account is first accessed.

Note that this not include newly loaded accounts, account and storage is considered warm if it is found in the State.

Implementations

impl JournaledState

pub fn new( spec: SpecId, warm_preloaded_addresses: HashSet<Address>, ) -> JournaledState

Create new JournaledState.

warm_preloaded_addresses is used to determine if address is considered warm loaded. In ordinary case this is precompile or beneficiary.

Note: This function will journal state after Spurious Dragon fork. And will not take into account if account is not existing or empty.

Note

pub fn state(&mut self) -> &mut HashMap<Address, Account>

Return reference to state.

pub fn set_spec_id(&mut self, spec: SpecId)

Sets SpecId.

pub fn touch(&mut self, address: &Address)

Mark account as touched as only touched accounts will be added to state. This is especially important for state clear where touched empty accounts needs to be removed from state.

pub fn clear(&mut self)

Clears the JournaledState. Preserving only the spec.

pub fn finalize(&mut self) -> (HashMap<Address, Account>, Vec<Log>)

Does cleanup and returns modified state.

This resets the JournaledState to its initial state in Self::new

pub fn account(&self, address: Address) -> &Account

Returns the loaded Account for the given address.

This assumes that the account has already been loaded.

Panics

Panics if the account has not been loaded and is missing from the state set.

pub fn depth(&self) -> u64

Returns call depth.

pub fn set_code(&mut self, address: Address, code: Bytecode)

use it only if you know that acc is warm Assume account is warm

pub fn inc_nonce(&mut self, address: Address) -> Option<u64>

pub fn transfer<DB>( &mut self, from: &Address, to: &Address, balance: Uint<256, 4>, db: &mut DB, ) -> Result<Option<InstructionResult>, EVMError<<DB as Database>::Error>>
where DB: Database,

Transfers balance from two accounts. Returns error if sender balance is not enough.

pub fn create_account_checkpoint( &mut self, caller: Address, address: Address, balance: Uint<256, 4>, spec_id: SpecId, ) -> Result<JournalCheckpoint, InstructionResult>

Create account or return false if collision is detected.

There are few steps done:

Make created account warm loaded (AccessList) and this should be done before subroutine checkpoint is created.

Check if there is collision of newly created account with existing one.

Mark created account as created.

Add fund to created account

Increment nonce of created account if SpuriousDragon is active

Decrease balance of caller account.

Panics

Panics if the caller is not loaded inside of the EVM state. This is should have been done inside create_inner.

pub fn checkpoint(&mut self) -> JournalCheckpoint

Makes a checkpoint that in case of Revert can bring back state to this point.

pub fn checkpoint_commit(&mut self)

Commit the checkpoint.

pub fn checkpoint_revert(&mut self, checkpoint: JournalCheckpoint)

Reverts all changes to state until given checkpoint.

pub fn selfdestruct<DB>( &mut self, address: Address, target: Address, db: &mut DB, ) -> Result<SelfDestructResult, EVMError<<DB as Database>::Error>>
where DB: Database,

Performances selfdestruct action. Transfers balance from address to target. Check if target exist/is_cold

Note: balance will be lost if address and target are the same BUT when current spec enables Cancun, this happens only when the account associated to address is created in the same tx

references:

https://github.com/ethereum/go-ethereum/blob/141cd425310b503c5678e674a8c3872cf46b7086/core/vm/instructions.go#L832-L833

https://github.com/ethereum/go-ethereum/blob/141cd425310b503c5678e674a8c3872cf46b7086/core/state/statedb.go#L449

https://eips.ethereum.org/EIPS/eip-6780

pub fn initial_account_load<DB>( &mut self, address: Address, slots: &[Uint<256, 4>], db: &mut DB, ) -> Result<&mut Account, EVMError<<DB as Database>::Error>>
where DB: Database,

Initial load of account. This load will not be tracked inside journal

pub fn load_account<DB>( &mut self, address: Address, db: &mut DB, ) -> Result<(&mut Account, bool), EVMError<<DB as Database>::Error>>
where DB: Database,

load account into memory. return if it is cold or warm accessed

pub fn load_account_exist<DB>( &mut self, address: Address, db: &mut DB, ) -> Result<LoadAccountResult, EVMError<<DB as Database>::Error>>
where DB: Database,

Load account from database to JournaledState.

Return boolean pair where first is is_cold second bool is_exists.

pub fn load_code<DB>( &mut self, address: Address, db: &mut DB, ) -> Result<(&mut Account, bool), EVMError<<DB as Database>::Error>>
where DB: Database,

Loads code.

pub fn sload<DB>( &mut self, address: Address, key: Uint<256, 4>, db: &mut DB, ) -> Result<(Uint<256, 4>, bool), EVMError<<DB as Database>::Error>>
where DB: Database,

Load storage slot

Panics

Panics if the account is not present in the state.

pub fn sstore<DB>( &mut self, address: Address, key: Uint<256, 4>, new: Uint<256, 4>, db: &mut DB, ) -> Result<SStoreResult, EVMError<<DB as Database>::Error>>
where DB: Database,

Stores storage slot. And returns (original,present,new) slot value.

Note:

account should already be present in our state.

pub fn tload(&mut self, address: Address, key: Uint<256, 4>) -> Uint<256, 4>

Read transient storage tied to the account.

EIP-1153: Transient storage opcodes

pub fn tstore(&mut self, address: Address, key: Uint<256, 4>, new: Uint<256, 4>)

Store transient storage tied to the account.

If values is different add entry to the journal so that old state can be reverted if that action is needed.

EIP-1153: Transient storage opcodes

pub fn log(&mut self, log: Log)

push log into subroutine