Struct cast::KmsClient

pub struct KmsClient {
    client: Client,
    region: Region,
}

A client for the KMS API.

Fields
client: Clientregion: Region
Implementations

impl KmsClient

pub fn new(region: Region) -> KmsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>( request_dispatcher: D, credentials_provider: P, region: Region ) -> KmsClient
where P: ProvideAwsCredentials + Send + Sync + 'static, D: DispatchSignedRequest + Send + Sync + 'static,

pub fn new_with_client(client: Client, region: Region) -> KmsClient