cast

Struct SimpleCast

pub struct SimpleCast;

Implementations

impl SimpleCast

pub fn max_int(s: &str) -> Result<String>

Returns the maximum value of the given integer type

Example

use alloy_primitives::{I256, U256};
use cast::SimpleCast;

assert_eq!(SimpleCast::max_int("uint256")?, U256::MAX.to_string());
assert_eq!(SimpleCast::max_int("int256")?, I256::MAX.to_string());
assert_eq!(SimpleCast::max_int("int32")?, i32::MAX.to_string());

pub fn min_int(s: &str) -> Result<String>

Returns the maximum value of the given integer type

Example

use alloy_primitives::{I256, U256};
use cast::SimpleCast;

assert_eq!(SimpleCast::min_int("uint256")?, "0");
assert_eq!(SimpleCast::min_int("int256")?, I256::MIN.to_string());
assert_eq!(SimpleCast::min_int("int32")?, i32::MIN.to_string());

pub(crate) fn max_min_int<const MAX: bool>(s: &str) -> Result<String>

pub fn from_utf8(s: &str) -> String

Converts UTF-8 text input to hex

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::from_utf8("yo"), "0x796f");
assert_eq!(Cast::from_utf8("Hello, World!"), "0x48656c6c6f2c20576f726c6421");
assert_eq!(Cast::from_utf8("TurboDappTools"), "0x547572626f44617070546f6f6c73");

pub fn to_utf8(s: &str) -> Result<String>

Converts hex input to UTF-8 text

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::to_utf8("0x796f")?, "yo");
assert_eq!(Cast::to_utf8("0x48656c6c6f2c20576f726c6421")?, "Hello, World!");
assert_eq!(Cast::to_utf8("0x547572626f44617070546f6f6c73")?, "TurboDappTools");
assert_eq!(Cast::to_utf8("0xe4bda0e5a5bd")?, "你好");

pub fn to_ascii(hex: &str) -> Result<String>

Converts hex data into text data

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::to_ascii("0x796f")?, "yo");
assert_eq!(Cast::to_ascii("48656c6c6f2c20576f726c6421")?, "Hello, World!");
assert_eq!(Cast::to_ascii("0x547572626f44617070546f6f6c73")?, "TurboDappTools");

pub fn from_fixed_point(value: &str, decimals: &str) -> Result<String>

Converts fixed point number into specified number of decimals

use alloy_primitives::U256;
use cast::SimpleCast as Cast;

assert_eq!(Cast::from_fixed_point("10", "0")?, "10");
assert_eq!(Cast::from_fixed_point("1.0", "1")?, "10");
assert_eq!(Cast::from_fixed_point("0.10", "2")?, "10");
assert_eq!(Cast::from_fixed_point("0.010", "3")?, "10");

pub fn to_fixed_point(value: &str, decimals: &str) -> Result<String>

Converts integers with specified decimals into fixed point numbers

Example

use alloy_primitives::U256;
use cast::SimpleCast as Cast;

assert_eq!(Cast::to_fixed_point("10", "0")?, "10.");
assert_eq!(Cast::to_fixed_point("10", "1")?, "1.0");
assert_eq!(Cast::to_fixed_point("10", "2")?, "0.10");
assert_eq!(Cast::to_fixed_point("10", "3")?, "0.010");

assert_eq!(Cast::to_fixed_point("-10", "0")?, "-10.");
assert_eq!(Cast::to_fixed_point("-10", "1")?, "-1.0");
assert_eq!(Cast::to_fixed_point("-10", "2")?, "-0.10");
assert_eq!(Cast::to_fixed_point("-10", "3")?, "-0.010");

pub fn concat_hex<T: AsRef<str>>(values: impl IntoIterator<Item = T>) -> String

Concatencates hex strings

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::concat_hex(["0x00", "0x01"]), "0x0001");
assert_eq!(Cast::concat_hex(["1", "2"]), "0x12");

pub fn to_uint256(value: &str) -> Result<String>

Converts a number into uint256 hex string with 0x prefix

Example

use cast::SimpleCast as Cast;

assert_eq!(
    Cast::to_uint256("100")?,
    "0x0000000000000000000000000000000000000000000000000000000000000064"
);
assert_eq!(
    Cast::to_uint256("192038293923")?,
    "0x0000000000000000000000000000000000000000000000000000002cb65fd1a3"
);
assert_eq!(
    Cast::to_uint256(
        "115792089237316195423570985008687907853269984665640564039457584007913129639935"
    )?,
    "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
);

pub fn to_int256(value: &str) -> Result<String>

Converts a number into int256 hex string with 0x prefix

Example

use cast::SimpleCast as Cast;

assert_eq!(
    Cast::to_int256("0")?,
    "0x0000000000000000000000000000000000000000000000000000000000000000"
);
assert_eq!(
    Cast::to_int256("100")?,
    "0x0000000000000000000000000000000000000000000000000000000000000064"
);
assert_eq!(
    Cast::to_int256("-100")?,
    "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff9c"
);
assert_eq!(
    Cast::to_int256("192038293923")?,
    "0x0000000000000000000000000000000000000000000000000000002cb65fd1a3"
);
assert_eq!(
    Cast::to_int256("-192038293923")?,
    "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffd349a02e5d"
);
assert_eq!(
    Cast::to_int256(
        "57896044618658097711785492504343953926634992332820282019728792003956564819967"
    )?,
    "0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
);
assert_eq!(
    Cast::to_int256(
        "-57896044618658097711785492504343953926634992332820282019728792003956564819968"
    )?,
    "0x8000000000000000000000000000000000000000000000000000000000000000"
);

pub fn to_unit(value: &str, unit: &str) -> Result<String>

Converts an eth amount into a specified unit

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::to_unit("1 wei", "wei")?, "1");
assert_eq!(Cast::to_unit("1", "wei")?, "1");
assert_eq!(Cast::to_unit("1ether", "wei")?, "1000000000000000000");

pub fn parse_units(value: &str, unit: u8) -> Result<String>

Convert a number into a uint with arbitrary decimals.

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::parse_units("1.0", 6)?, "1000000"); // USDC (6 decimals)
assert_eq!(Cast::parse_units("2.5", 6)?, "2500000");
assert_eq!(Cast::parse_units("1.0", 12)?, "1000000000000"); // 12 decimals
assert_eq!(Cast::parse_units("1.23", 3)?, "1230"); // 3 decimals

pub fn format_units(value: &str, unit: u8) -> Result<String>

Format a number from smallest unit to decimal with arbitrary decimals.

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::format_units("1000000", 6)?, "1"); // USDC (6 decimals)
assert_eq!(Cast::format_units("2500000", 6)?, "2.500000");
assert_eq!(Cast::format_units("1000000000000", 12)?, "1"); // 12 decimals
assert_eq!(Cast::format_units("1230", 3)?, "1.230"); // 3 decimals

pub(crate) fn format_unit_as_string(value: U256, unit: Unit) -> String

pub fn from_wei(value: &str, unit: &str) -> Result<String>

Converts wei into an eth amount

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::from_wei("1", "gwei")?, "0.000000001");
assert_eq!(Cast::from_wei("12340000005", "gwei")?, "12.340000005");
assert_eq!(Cast::from_wei("10", "ether")?, "0.000000000000000010");
assert_eq!(Cast::from_wei("100", "eth")?, "0.000000000000000100");
assert_eq!(Cast::from_wei("17", "ether")?, "0.000000000000000017");

pub fn to_wei(value: &str, unit: &str) -> Result<String>

Converts an eth amount into wei

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::to_wei("100", "gwei")?, "100000000000");
assert_eq!(Cast::to_wei("100", "eth")?, "100000000000000000000");
assert_eq!(Cast::to_wei("1000", "ether")?, "1000000000000000000000");

pub fn from_rlp(value: impl AsRef<str>, as_int: bool) -> Result<String>

Examples

use cast::SimpleCast as Cast;

assert_eq!(Cast::from_rlp("0xc0", false).unwrap(), "[]");
assert_eq!(Cast::from_rlp("0x0f", false).unwrap(), "\"0x0f\"");
assert_eq!(Cast::from_rlp("0x33", false).unwrap(), "\"0x33\"");
assert_eq!(Cast::from_rlp("0xc161", false).unwrap(), "[\"0x61\"]");
assert_eq!(Cast::from_rlp("820002", true).is_err(), true);
assert_eq!(Cast::from_rlp("820002", false).unwrap(), "\"0x0002\"");
assert_eq!(Cast::from_rlp("00", true).is_err(), true);
assert_eq!(Cast::from_rlp("00", false).unwrap(), "\"0x00\"");

pub fn to_rlp(value: &str) -> Result<String>

Encodes hex data or list of hex data to hexadecimal rlp

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::to_rlp("[]").unwrap(), "0xc0".to_string());
assert_eq!(Cast::to_rlp("0x22").unwrap(), "0x22".to_string());
assert_eq!(Cast::to_rlp("[\"0x61\"]",).unwrap(), "0xc161".to_string());
assert_eq!(Cast::to_rlp("[\"0xf1\", \"f2\"]").unwrap(), "0xc481f181f2".to_string());

pub fn to_base( value: &str, base_in: Option<&str>, base_out: &str, ) -> Result<String>

Converts a number of one base to another

Example

use alloy_primitives::I256;
use cast::SimpleCast as Cast;

assert_eq!(Cast::to_base("100", Some("10"), "16")?, "0x64");
assert_eq!(Cast::to_base("100", Some("10"), "oct")?, "0o144");
assert_eq!(Cast::to_base("100", Some("10"), "binary")?, "0b1100100");

assert_eq!(Cast::to_base("0xffffffffffffffff", None, "10")?, u64::MAX.to_string());
assert_eq!(
    Cast::to_base("0xffffffffffffffffffffffffffffffff", None, "dec")?,
    u128::MAX.to_string()
);
// U256::MAX overflows as internally it is being parsed as I256
assert_eq!(
    Cast::to_base(
        "0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
        None,
        "decimal"
    )?,
    I256::MAX.to_string()
);

pub fn to_bytes32(s: &str) -> Result<String>

Converts hexdata into bytes32 value

Example

use cast::SimpleCast as Cast;

let bytes = Cast::to_bytes32("1234")?;
assert_eq!(bytes, "0x1234000000000000000000000000000000000000000000000000000000000000");

let bytes = Cast::to_bytes32("0x1234")?;
assert_eq!(bytes, "0x1234000000000000000000000000000000000000000000000000000000000000");

let err = Cast::to_bytes32("0x123400000000000000000000000000000000000000000000000000000000000011").unwrap_err();
assert_eq!(err.to_string(), "string >32 bytes");

pub fn format_bytes32_string(s: &str) -> Result<String>

Encodes string into bytes32 value

pub fn pad(s: &str, right: bool, len: usize) -> Result<String>

Pads hex data to a specified length

Example

use cast::SimpleCast as Cast;

let padded = Cast::pad("abcd", true, 20)?;
assert_eq!(padded, "0xabcd000000000000000000000000000000000000");

let padded = Cast::pad("abcd", false, 20)?;
assert_eq!(padded, "0x000000000000000000000000000000000000abcd");

let padded = Cast::pad("0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2", true, 32)?;
assert_eq!(padded, "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2000000000000000000000000");

let padded = Cast::pad("0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2", false, 32)?;
assert_eq!(padded, "0x000000000000000000000000C02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2");

let err = Cast::pad("1234", false, 1).unwrap_err();
assert_eq!(err.to_string(), "input length exceeds target length");

let err = Cast::pad("foobar", false, 32).unwrap_err();
assert_eq!(err.to_string(), "input is not a valid hex");

pub fn parse_bytes32_string(s: &str) -> Result<String>

Decodes string from bytes32 value

pub fn parse_bytes32_address(s: &str) -> Result<String>

Decodes checksummed address from bytes32 value

pub fn abi_decode( sig: &str, calldata: &str, input: bool, ) -> Result<Vec<DynSolValue>>

Decodes abi-encoded hex input or output

When input=true, calldata string MUST not be prefixed with function selector

Example

use cast::SimpleCast as Cast;
use alloy_primitives::hex;

    // Passing `input = false` will decode the data as the output type.
    // The input data types and the full function sig are ignored, i.e.
    // you could also pass `balanceOf()(uint256)` and it'd still work.
    let data = "0x0000000000000000000000000000000000000000000000000000000000000001";
    let sig = "balanceOf(address, uint256)(uint256)";
    let decoded = Cast::abi_decode(sig, data, false)?[0].as_uint().unwrap().0.to_string();
    assert_eq!(decoded, "1");

    // Passing `input = true` will decode the data with the input function signature.
    // We exclude the "prefixed" function selector from the data field (the first 4 bytes).
    let data = "0x0000000000000000000000008dbd1b711dc621e1404633da156fcc779e1c6f3e000000000000000000000000d9f3c9cc99548bf3b44a43e0a2d07399eb918adc000000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000000";
    let sig = "safeTransferFrom(address, address, uint256, uint256, bytes)";
    let decoded = Cast::abi_decode(sig, data, true)?;
    let decoded = [
        decoded[0].as_address().unwrap().to_string().to_lowercase(),
        decoded[1].as_address().unwrap().to_string().to_lowercase(),
        decoded[2].as_uint().unwrap().0.to_string(),
        decoded[3].as_uint().unwrap().0.to_string(),
        hex::encode(decoded[4].as_bytes().unwrap())
    ]
    .into_iter()
    .collect::<Vec<_>>();

    assert_eq!(
        decoded,
        vec!["0x8dbd1b711dc621e1404633da156fcc779e1c6f3e", "0xd9f3c9cc99548bf3b44a43e0a2d07399eb918adc", "42", "1", ""]
    );

pub fn calldata_decode( sig: &str, calldata: &str, input: bool, ) -> Result<Vec<DynSolValue>>

Decodes calldata-encoded hex input or output

Similar to abi_decode, but calldata string MUST be prefixed with function selector

Example

use cast::SimpleCast as Cast;
use alloy_primitives::hex;

// Passing `input = false` will decode the data as the output type.
// The input data types and the full function sig are ignored, i.e.
// you could also pass `balanceOf()(uint256)` and it'd still work.
let data = "0x0000000000000000000000000000000000000000000000000000000000000001";
let sig = "balanceOf(address, uint256)(uint256)";
let decoded = Cast::calldata_decode(sig, data, false)?[0].as_uint().unwrap().0.to_string();
assert_eq!(decoded, "1");

    // Passing `input = true` will decode the data with the input function signature.
    let data = "0xf242432a0000000000000000000000008dbd1b711dc621e1404633da156fcc779e1c6f3e000000000000000000000000d9f3c9cc99548bf3b44a43e0a2d07399eb918adc000000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000000";
    let sig = "safeTransferFrom(address, address, uint256, uint256, bytes)";
    let decoded = Cast::calldata_decode(sig, data, true)?;
    let decoded = [
        decoded[0].as_address().unwrap().to_string().to_lowercase(),
        decoded[1].as_address().unwrap().to_string().to_lowercase(),
        decoded[2].as_uint().unwrap().0.to_string(),
        decoded[3].as_uint().unwrap().0.to_string(),
        hex::encode(decoded[4].as_bytes().unwrap()),
   ]
   .into_iter()
   .collect::<Vec<_>>();
    assert_eq!(
        decoded,
        vec!["0x8dbd1b711dc621e1404633da156fcc779e1c6f3e", "0xd9f3c9cc99548bf3b44a43e0a2d07399eb918adc", "42", "1", ""]
    );

pub fn abi_encode(sig: &str, args: &[impl AsRef<str>]) -> Result<String>

Performs ABI encoding based off of the function signature. Does not include the function selector in the result.

Example

use cast::SimpleCast as Cast;

assert_eq!(
    "0x0000000000000000000000000000000000000000000000000000000000000001",
    Cast::abi_encode("f(uint a)", &["1"]).unwrap().as_str()
);
assert_eq!(
    "0x0000000000000000000000000000000000000000000000000000000000000001",
    Cast::abi_encode("constructor(uint a)", &["1"]).unwrap().as_str()
);

pub fn abi_encode_packed(sig: &str, args: &[impl AsRef<str>]) -> Result<String>

Performs packed ABI encoding based off of the function signature or tuple.

Examplez

use cast::SimpleCast as Cast;

assert_eq!(
    "0x0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000012c00000000000000c8",
    Cast::abi_encode_packed("(uint128[] a, uint64 b)", &["[100, 300]", "200"]).unwrap().as_str()
);

assert_eq!(
    "0x8dbd1b711dc621e1404633da156fcc779e1c6f3e68656c6c6f20776f726c64",
    Cast::abi_encode_packed("foo(address a, string b)", &["0x8dbd1b711dc621e1404633da156fcc779e1c6f3e", "hello world"]).unwrap().as_str()
);

pub fn calldata_encode( sig: impl AsRef<str>, args: &[impl AsRef<str>], ) -> Result<String>

Performs ABI encoding to produce the hexadecimal calldata with the given arguments.

Example

use cast::SimpleCast as Cast;

assert_eq!(
    "0xb3de648b0000000000000000000000000000000000000000000000000000000000000001",
    Cast::calldata_encode("f(uint256 a)", &["1"]).unwrap().as_str()
);

pub fn index(key_type: &str, key: &str, slot_number: &str) -> Result<String>

Returns the slot number for a given mapping key and slot.

Given mapping(k => v) m, for a key k the slot number of its associated v is keccak256(concat(h(k), p)), where h is the padding function for k’s type, and p is slot number of the mapping m.

See the Solidity documentation for more details.

Example

// Value types.
assert_eq!(
    Cast::index("address", "0xD0074F4E6490ae3f888d1d4f7E3E43326bD3f0f5", "2").unwrap().as_str(),
    "0x9525a448a9000053a4d151336329d6563b7e80b24f8e628e95527f218e8ab5fb"
);
assert_eq!(
    Cast::index("uint256", "42", "6").unwrap().as_str(),
    "0xfc808b0f31a1e6b9cf25ff6289feae9b51017b392cc8e25620a94a38dcdafcc1"
);

// Strings and byte arrays.
assert_eq!(
    Cast::index("string", "hello", "1").unwrap().as_str(),
    "0x8404bb4d805e9ca2bd5dd5c43a107e935c8ec393caa7851b353b3192cd5379ae"
);

pub fn keccak(data: &str) -> Result<String>

Keccak-256 hashes arbitrary data

Example

use cast::SimpleCast as Cast;

assert_eq!(
    Cast::keccak("foo")?,
    "0x41b1a0649752af1b28b3dc29a1556eee781e4a4c3a1f7f53f90fa834de098c4d"
);
assert_eq!(
    Cast::keccak("123abc")?,
    "0xb1f1c74a1ba56f07a892ea1110a39349d40f66ca01d245e704621033cb7046a4"
);
assert_eq!(
    Cast::keccak("0x12")?,
    "0x5fa2358263196dbbf23d1ca7a509451f7a2f64c15837bfbb81298b1e3e24e4fa"
);
assert_eq!(
    Cast::keccak("12")?,
    "0x7f8b6b088b6d74c2852fc86c796dca07b44eed6fb3daf5e6b59f7c364db14528"
);

pub fn left_shift( value: &str, bits: &str, base_in: Option<&str>, base_out: &str, ) -> Result<String>

Performs the left shift operation (<<) on a number

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::left_shift("16", "10", Some("10"), "hex")?, "0x4000");
assert_eq!(Cast::left_shift("255", "16", Some("dec"), "hex")?, "0xff0000");
assert_eq!(Cast::left_shift("0xff", "16", None, "hex")?, "0xff0000");

pub fn right_shift( value: &str, bits: &str, base_in: Option<&str>, base_out: &str, ) -> Result<String>

Performs the right shift operation (>>) on a number

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::right_shift("0x4000", "10", None, "dec")?, "16");
assert_eq!(Cast::right_shift("16711680", "16", Some("10"), "hex")?, "0xff");
assert_eq!(Cast::right_shift("0xff0000", "16", None, "hex")?, "0xff");

pub async fn etherscan_source( chain: Chain, contract_address: String, etherscan_api_key: Option<String>, explorer_api_url: Option<String>, explorer_url: Option<String>, ) -> Result<String>

Fetches source code of verified contracts from etherscan.

Example

assert_eq!(
    "/*
            - Bytecode Verification performed was compared on second iteration -
            This file is part of the DAO.....",
    Cast::etherscan_source(
        NamedChain::Mainnet.into(),
        "0xBB9bc244D798123fDe783fCc1C72d3Bb8C189413".to_string(),
        Some("<etherscan_api_key>".to_string()),
        None,
        None
    )
    .await
    .unwrap()
    .as_str()
);

pub async fn expand_etherscan_source_to_directory( chain: Chain, contract_address: String, etherscan_api_key: Option<String>, output_directory: PathBuf, explorer_api_url: Option<String>, explorer_url: Option<String>, ) -> Result<()>

Fetches the source code of verified contracts from etherscan and expands the resulting files to a directory for easy perusal.

Example

Cast::expand_etherscan_source_to_directory(
    NamedChain::Mainnet.into(),
    "0xBB9bc244D798123fDe783fCc1C72d3Bb8C189413".to_string(),
    Some("<etherscan_api_key>".to_string()),
    PathBuf::from("output_dir"),
    None,
    None,
)
.await?;

pub async fn etherscan_source_flatten( chain: Chain, contract_address: String, etherscan_api_key: Option<String>, output_path: Option<PathBuf>, explorer_api_url: Option<String>, explorer_url: Option<String>, ) -> Result<()>

Fetches the source code of verified contracts from etherscan, flattens it and writes it to the given path or stdout.

pub fn disassemble(code: &[u8]) -> Result<String>

Disassembles hex encoded bytecode into individual / human readable opcodes

Example

use alloy_primitives::hex;
use cast::SimpleCast as Cast;

let bytecode = "0x608060405260043610603f57600035";
let opcodes = Cast::disassemble(&hex::decode(bytecode)?)?;
println!("{}", opcodes);

pub fn get_selector( signature: &str, optimize: usize, ) -> Result<(String, String)>

Gets the selector for a given function signature Optimizes if the optimize parameter is set to a number of leading zeroes

Example

use cast::SimpleCast as Cast;

assert_eq!(Cast::get_selector("foo(address,uint256)", 0)?.0, String::from("0xbd0d639f"));

pub fn extract_functions( bytecode: &str, ) -> Result<Vec<(Selector, String, &str)>>

Extracts function selectors, arguments and state mutability from bytecode

Example

use alloy_primitives::fixed_bytes;
use cast::SimpleCast as Cast;

let bytecode = "6080604052348015600e575f80fd5b50600436106026575f3560e01c80632125b65b14602a575b5f80fd5b603a6035366004603c565b505050565b005b5f805f60608486031215604d575f80fd5b833563ffffffff81168114605f575f80fd5b925060208401356001600160a01b03811681146079575f80fd5b915060408401356001600160e01b03811681146093575f80fd5b80915050925092509256";
let functions = Cast::extract_functions(bytecode)?;
assert_eq!(functions, vec![(fixed_bytes!("0x2125b65b"), "uint32,address,uint224".to_string(), "pure")]);

pub fn decode_raw_transaction(tx: &str) -> Result<TxEnvelope>

Decodes a raw EIP2718 transaction payload Returns details about the typed transaction and ECSDA signature components

Example

use cast::SimpleCast as Cast;

let tx = "0x02f8f582a86a82058d8459682f008508351050808303fd84948e42f2f4101563bf679975178e880fd87d3efd4e80b884659ac74b00000000000000000000000080f0c1c49891dcfdd40b6e0f960f84e6042bcb6f000000000000000000000000b97ef9ef8734c71904d8002f8b6bc66dd9c48a6e00000000000000000000000000000000000000000000000000000000007ff4e20000000000000000000000000000000000000000000000000000000000000064c001a05d429597befe2835396206781b199122f2e8297327ed4a05483339e7a8b2022aa04c23a7f70fb29dda1b4ee342fb10a625e9b8ddc6a603fb4e170d4f6f37700cb8";
let tx_envelope = Cast::decode_raw_transaction(&tx)?;