Module cast::traces::identifier

Call trace address identifiers.

Identifiers figure out what ABIs and labels belong to all the addresses of the trace.

Structs

AddressIdentity

An address identity

EtherscanIdentifier

A trace identifier that tries to identify addresses using Etherscan.

LocalTraceIdentifier

A trace identifier that tries to identify addresses using local contracts.

SignaturesIdentifier

An identifier that tries to identify functions and events using signatures found at https://openchain.xyz.

TraceIdentifiers

A collection of trace identifiers.

Traits

TraceIdentifier

Trace identifiers figure out what ABIs and labels belong to all the addresses of the trace.

Type Aliases

SingleSignaturesIdentifier