Trait cast::InspectorExt

pub trait InspectorExt<DB>: Inspector<DB>
where
    DB: Database,
{
    // Provided method
    fn should_use_create2_factory(
        &mut self,
        _context: &mut EvmContext<DB>,
        _inputs: &mut CreateInputs,
    ) -> bool { ... }
}

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

Provided Methods

fn should_use_create2_factory( &mut self, _context: &mut EvmContext<DB>, _inputs: &mut CreateInputs, ) -> bool

Determines whether the DEFAULT_CREATE2_DEPLOYER should be used for a CREATE2 frame.

If this function returns true, we’ll replace CREATE2 frame with a CALL frame to CREATE2 factory.