cast

Module tx

Structs

CastTxBuilder

Builder type constructing [TransactionRequest] from cast send/mktx inputs.

InitState

Initial state.

InputState

State with known input for the transaction.

ToState

State with known TxKind.

Enums

SenderKind

Different sender kinds used by CastTxBuilder.

Functions

decode_execution_revert 🔒

Helper function that tries to decode custom error name and inputs from error payload data.

validate_from_address

Prevents a misconfigured hwlib from sending a transaction that defies user-specified –from