cast::utils

Type Alias StateChangeset

pub type StateChangeset = HashMap<Address, Account, RandomState>;

EVM State is a mapping from addresses to accounts.

Aliased Type

pub struct StateChangeset { /* private fields */ }

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 40 bytes