Crate cast

Modules

backend

Foundry’s main executor backend abstraction and implementation.

base

constants

coverage

foundry-evm-coverage

debug

decode

Various utilities to decode test results.

errors

Errors for this crate

executors

EVM executor abstractions, which can execute calls.

fork

fuzz

foundry-evm-fuzz

inspectors

EVM inspectors.

opts

revm

Revm is a Rust EVM implementation.

rlp_converter 🔒

traces

foundry-evm-traces

tx 🔒

utils

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch.

Structs

AwsChainProvider

Provides AWS credentials from multiple possible sources using a priority order.

AwsClient

Re-usable logic for all clients.

AwsHttpClient

Http client for use with AWS services.

Cast

InterfaceSource

KmsClient

A client for the KMS API.

SimpleCast

TxBuilder

Transaction builder

Enums

AbiPath

AwsRegion

An AWS region.

Functions

strip_0x 🔒

Type Aliases

TxBuilderOutput