Crate cast

cast

Cast is a command-line tool for performing Ethereum RPC calls. You can make smart contract calls, send transactions, or retrieve any type of chain data - all from your command-line!

For more information, see the 📖 Foundry Book (Cast Guide).

Modules

IERC20

Module containing a contract’s types and functions.

backend

Foundry’s main executor backend abstraction and implementation.

base

constants

coverage

foundry-evm-coverage

decode

Various utilities to decode test results.

errors

Errors for this crate

executors

EVM executor abstractions, which can execute calls.

fork

fuzz

foundry-evm-fuzz

inspectors

EVM inspectors.

opts

revm

Revm is a Rust EVM implementation.

rlp_converter 🔒

traces

foundry-evm-traces

utils

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch.

Structs

Cast

SimpleCast

Traits

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

Functions

strip_0x 🔒