Crate cast

Checked scalar casting

Examples

extern crate cast;

use cast::From;

// Infallible operations, like integer promotion, are equivalent to an `as` call
assert_eq!(u16::from(0u8), 0u16);

// Everything else will return an `Option` depending on the success of the operation
assert_eq!(u8::from(256u16), None);  // Overflow
assert_eq!(u8::from(-1i8), None);  // Underflow
assert_eq!(u8::from(0. / 0.), None);  // NaN
assert_eq!(u8::from(127i8), Some(127u8));  // OK

from_ vs from

Importing cast::From shadows std::convert::From, so you can no longer write code like Cow::from("Hello"). The logical solution is to rename the cast::From import:

extern crate cast;

use std::borrow::Cow;

// don't shadow `std::convert::From`
use cast::From as _0;

Cow::from("Hello");
u16::from(0u8);  //~ error: multiple applicable methods in scope

But then you'll hit this bug. The workaround for these cases where you want to use both convert::From and cast::From in the same scope is to use the from_ method to refer to the latter trait.

extern crate cast;

use std::borrow::Cow;

// don't shadow `std::convert::From`
use cast::From as _0;

Cow::from("Hello");
u16::from_(0u8);

Traits

From
The "cast from" operation