cast
Module fork

Modules

database

A revm database that forks off a remote client

Structs

CreateFork

Represents a fork of a remote chain whose data is available only via the url endpoint.

ForkId

The unique identifier for a specific fork, this could be the name of the network a custom descriptive name.

MultiFork

The Sender half of multi fork pair. Can send requests to the MultiForkHandler to create forks.

MultiForkHandler

The type that manages connections in the background.

Functions

environment

Initializes a REVM block environment based on a forked ethereum provider.