Module cast::fuzz::strategies

Structs

IntStrategy

Value tree for signed ints (up to int256). The strategy combines 3 different strategies, each assigned a specific weight:

UintStrategy

Value tree for unsigned ints (up to uint256). The strategy combines 3 different strategies, each assigned a specific weight:

Functions

build_initial_state

Builds the initial EvmFuzzState from a database.

collect_created_contracts

Collects all created contracts from a StateChangeset which haven’t been discovered yet. Stores them at targeted_contracts and created_contracts.

collect_state_from_call

Collects state changes from a StateChangeset and logs into an EvmFuzzState according to the given [FuzzDictionaryConfig].

fuzz_calldata

Given a function, it returns a strategy which generates valid calldata for that function’s input types.

fuzz_calldata_from_state

Given a function and some state, it returns a strategy which generated valid calldata for the given function’s input types, based on state taken from the EVM.

fuzz_contract_with_calldata

Given a function, it returns a proptest strategy which generates valid abi-encoded calldata for that function’s input types.

fuzz_param

Given a parameter type, returns a strategy for generating values for that type.

fuzz_param_from_state

Given a parameter type, returns a strategy for generating values for that type, given some EVM fuzz state.

invariant_strat

Creates the invariant strategy.

override_call_strat

Given a target address, we generate random calldata.

Type Aliases

EvmFuzzState

A set of arbitrary 32 byte data from the VM used to generate values for the strategy.