cast::inspectors

Crate cheatcodes

foundry-cheatcodes

Foundry cheatcodes implementations.

Modules

spec

Cheatcode specification for Foundry.

Structs

BroadcastableTransaction

Helps collecting transactions from different forks.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

CheatsConfig

Additional, configurable context the Cheatcodes inspector has access to

CheatsCtxt

The cheatcode context, used in Cheatcode.

Error

Error thrown by cheatcodes.

Wallets

Cloneable wrapper around WalletsInner.

WalletsInner

Contains context for wallet management.

Enums

ErrorKind

Kind of cheatcode errors.

ForgeContext

Forge execution contexts.

Traits

CheatcodeDef

Cheatcode definition trait. Implemented by all Vm functions.

CheatcodesExecutor

Helper trait for obtaining complete revm::Inspector instance from mutable reference to Cheatcodes.

Functions

set_execution_context

Set forge command current execution context for the duration of the program. Execution context is immutable, subsequent calls of this function won’t change the context.

Type Aliases

BroadcastableTransactions

List of transactions that can be broadcasted.

ExpectedCallTracker

Tracks the expected calls per address.

Result

Cheatcode result type.