Trait cast::inspectors::cheatcodes::CheatcodesExecutor

pub trait CheatcodesExecutor {
    // Required method
    fn get_inspector<'a, DB>(
        &'a mut self,
        cheats: &'a mut Cheatcodes,
    ) -> impl InspectorExt<DB> + 'a
       where DB: DatabaseExt;

    // Provided methods
    fn with_evm<DB, F, O>(
        &mut self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
        f: F,
    ) -> Result<O, EVMError<<DB as Database>::Error>>
       where DB: DatabaseExt,
             F: for<'a, 'b> FnOnce(&mut Evm<'_, &'b mut dyn InspectorExt<&'a mut dyn DatabaseExt<Error = DatabaseError>>, &'a mut dyn DatabaseExt<Error = DatabaseError>>) -> Result<O, EVMError<<DB as Database>::Error>> { ... }
    fn exec_create<DB>(
        &mut self,
        inputs: CreateInputs,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
    ) -> Result<CreateOutcome, EVMError<<DB as Database>::Error>>
       where DB: DatabaseExt { ... }
    fn console_log<DB>(
        &mut self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
        message: String,
    ) -> Result<(), EVMError<<DB as Database>::Error>>
       where DB: DatabaseExt { ... }
}

Helper trait for obtaining complete revm::Inspector instance from mutable reference to Cheatcodes.

This is needed for cases when inspector itself needs mutable access to Cheatcodes state and allows us to correctly execute arbitrary EVM frames from inside cheatcode implementations.

Required Methods

fn get_inspector<'a, DB>( &'a mut self, cheats: &'a mut Cheatcodes, ) -> impl InspectorExt<DB> + 'a
where DB: DatabaseExt,

Core trait method accepting mutable reference to Cheatcodes and returning revm::Inspector.

Provided Methods

fn with_evm<DB, F, O>( &mut self, ccx: &mut CheatsCtxt<'_, '_, DB>, f: F, ) -> Result<O, EVMError<<DB as Database>::Error>>
where DB: DatabaseExt, F: for<'a, 'b> FnOnce(&mut Evm<'_, &'b mut dyn InspectorExt<&'a mut dyn DatabaseExt<Error = DatabaseError>>, &'a mut dyn DatabaseExt<Error = DatabaseError>>) -> Result<O, EVMError<<DB as Database>::Error>>,

Constructs revm::Evm and runs a given closure with it.

fn exec_create<DB>( &mut self, inputs: CreateInputs, ccx: &mut CheatsCtxt<'_, '_, DB>, ) -> Result<CreateOutcome, EVMError<<DB as Database>::Error>>
where DB: DatabaseExt,

Obtains revm::Evm instance and executes the given CREATE frame.

fn console_log<DB>( &mut self, ccx: &mut CheatsCtxt<'_, '_, DB>, message: String, ) -> Result<(), EVMError<<DB as Database>::Error>>
where DB: DatabaseExt,

Object Safety

This trait is not object safe.