Crate cast::revm

Modules

db

Database implementations.

handler

inspectors

Inspector implementations.

interpreter

revm-interpreter

optimism

Optimism-specific constants, types, and helpers.

precompile

revm-precompile

Structs

CacheState

Cache state contains both modified and original values.

EVM

Struct that takes Database and enabled transact to update state directly to database. additionally it allows user to set all environment parameters.

EVMData

EVM Data contains all the data that EVM needs to execute.

EVMImpl

Handler

Handler acts as a proxy and allow to define different behavior for different sections of the code. This allows nice integration of different chains or to disable some mainnet behavior.

JournalCheckpoint

SubRoutine checkpoint that will help us to go back from this

JournaledState

JournalState is internal EVM state that is used to contain state and track changes to that state. It contains journal of changes that happened to state so that they can be reverted.

L1BlockInfo

L1 block info

State

State of blockchain.

StateBuilder

Allows building of State and initializing it with different options.

TransitionAccount

Account Created when EVM state is merged to cache state. And it is sent to Block state.

TransitionState

Enums

JournalEntry

Journal entries that are used to track changes to the state and are used to revert it.

Constants

BASE_FEE_RECIPIENT

The address of the base fee recipient.

CALL_STACK_LIMIT

EVM call stack limit.

L1_BLOCK_CONTRACT

The address of the L1Block contract.

L1_FEE_RECIPIENT

The address of L1 fee recipient.

Traits

Database

EVM database interface.

DatabaseCommit

EVM database commit interface.

DatabaseRef

EVM database interface.

Inspector

EVM Interpreter callbacks.

Transact

EVM transaction interface.

Functions

evm_inner

inspector_instruction

Outer closure that calls Inspector for every instruction.

new

Type Aliases

DBBox

Database boxed with a lifetime and Send.

DummyStateDB

InMemoryDB

A Database implementation that stores all state changes in memory.

StateDBBox

More constrained version of State that uses Boxed database with a lifetime.