cast::revm

Module db

Database implementations.

Modules

components

Database that is split on State and BlockHash traits.

emptydb

in_memory_db

states

Structs

AccountRevert

Assumption is that Revert can return full state from any future state to any past state.

BenchmarkDB

Custom benchmarking DB that only has account info for the zero address.

BundleAccount

Account information focused on creating of database changesets and Reverts.

BundleState

Bundle state contain only values that got changed

CacheDB

A Database implementation that stores all state changes in memory.

CacheState

Cache state contains both modified and original values.

DatabaseComponents

DbAccount

EmptyDBTyped

An empty database that always returns default values when queried.

PlainAccount

State

State of blockchain.

StateBuilder

Allows building of State and initializing it with different options.

TransitionAccount

Account Created when EVM state is merged to cache state. And it is sent to Block state.

TransitionState

WrapDatabaseRef

Wraps a DatabaseRef to provide a Database implementation.

Enums

AccountState

AccountStatus

AccountStatus represents the various states an account can be in after being loaded from the database.

DatabaseComponentError

OriginalValuesKnown

Option for BundleState when converting it to the plain state.

RevertToSlot

So storage can have multiple types:

Traits

BlockHash

BlockHashRef

Database

EVM database interface.

DatabaseCommit

EVM database commit interface.

DatabaseRef

EVM database interface.

StateRef

Type Aliases

DBBox

Database boxed with a lifetime and Send.

EmptyDB

An empty database that always returns default values when queried.

InMemoryDB

A Database implementation that stores all state changes in memory.

StateDBBox

More constrained version of State that uses Boxed database with a lifetime.

StorageWithOriginalValues

This storage represent values that are before block changed.