cast::revm::interpreter
Module interpreter

Modules

analysis

serdeserde

Structs

Contract

EVM contract information.

Interpreter

EVM bytecode interpreter.

InterpreterResult

The result of an interpreter operation.

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack with STACK_LIMIT capacity of words.

Constants

EMPTY_SHARED_MEMORY

Empty shared memory.

STACK_LIMIT

EVM interpreter stack limit.

Functions

num_words

Returns number of words what would fit to provided number of bytes, i.e. it rounds up the number bytes to number of words.

resize_memory

Resize the memory to the new size. Returns whether the gas was enough to resize the memory.