cast::revm::interpreter

Module interpreter

Core interpreter implementation for EVM bytecode execution. Core interpreter implementation and components.

Modules

ext_bytecode

Extended bytecode functionality.

Structs

EthInterpreter

Default types for Ethereum interpreter.

ExtBytecode

Extended bytecode structure that wraps base bytecode with additional execution metadata.

InputsImpl

Inputs for the interpreter that are used for execution of the call.

Interpreter

Main interpreter structure that contains all components defined in InterpreterTypes.

InterpreterResult

The result of an interpreter operation.

ReturnDataImpl

Default implementation of return data storage for the interpreter.

RuntimeFlags

Runtime flags that control interpreter execution behavior.

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack with STACK_LIMIT capacity of words.

Constants

STACK_LIMIT

EVM interpreter stack limit.

Functions

num_words

Returns number of words what would fit to provided number of bytes, i.e. it rounds up the number bytes to number of words.

resize_memory

Performs EVM memory resize.