cast::revm

Module optimism

Available on crate feature optimism only.

Optimism-specific constants, types, and helpers.

Structs

L1BlockInfo

L1 block info

Constants

BASE_FEE_RECIPIENT

The address of the base fee recipient.

L1_BLOCK_CONTRACT

The address of the L1Block contract.

L1_FEE_RECIPIENT

The address of L1 fee recipient.

OPERATOR_FEE_RECIPIENT

The address of the operator fee recipient.

Functions

clear

Clears cache OP l1 value.

deduct_caller

Deduct max balance from caller

end

Optimism end handle changes output if the transaction is a deposit transaction. Deposit transaction can’t be reverted and is always successful.

last_frame_return

Handle output of the transaction

load_precompiles

Load precompiles for Optimism chain.

optimism_handle_register

output

Main return handle, returns the output of the transaction.

refund

Record Eip-7702 refund and calculate final refund.

reimburse_caller

Reimburse the transaction caller.

reward_beneficiary

Reward beneficiary with gas fee.

validate_env

Validate environment for the Optimism chain.

validate_tx_against_state

Don not perform any extra validation for deposit transactions, they are pre-verified on L1.