Module cast::revm::precompile::secp256r1

Available on crate feature secp256r1 only.

EIP-7212 secp256r1 Precompile

This module implements the EIP-7212 precompile for secp256r1 curve support.

The main purpose of this precompile is to verify ECDSA signatures that use the secp256r1, or P256 elliptic curve. The P256VERIFY const represents the implementation of this precompile, with the address that it is currently deployed at.

Constants

P256VERIFY

EIP-7212 secp256r1 precompile.

Functions

p256_verify

secp256r1 precompile logic. It takes the input bytes sent to the precompile and the gas limit. The output represents the result of verifying the secp256r1 signature of the input.

precompiles

Returns the secp256r1 precompile with its address.

verify_impl

Returns Some(()) if the signature included in the input byte slice is valid, None otherwise.