Trait cast::revm::Inspector

pub trait Inspector<DB>
where
    DB: Database,
{
    // Provided methods
    fn initialize_interp(
        &mut self,
        interp: &mut Interpreter<'_>,
        data: &mut EVMData<'_, DB>
    ) { ... }
    fn step(&mut self, interp: &mut Interpreter<'_>, data: &mut EVMData<'_, DB>) { ... }
    fn log(
        &mut self,
        evm_data: &mut EVMData<'_, DB>,
        address: &Address,
        topics: &[FixedBytes<32>],
        data: &Bytes
    ) { ... }
    fn step_end(
        &mut self,
        interp: &mut Interpreter<'_>,
        data: &mut EVMData<'_, DB>
    ) { ... }
    fn call(
        &mut self,
        data: &mut EVMData<'_, DB>,
        inputs: &mut CallInputs
    ) -> (InstructionResult, Gas, Bytes) { ... }
    fn call_end(
        &mut self,
        data: &mut EVMData<'_, DB>,
        inputs: &CallInputs,
        remaining_gas: Gas,
        ret: InstructionResult,
        out: Bytes
    ) -> (InstructionResult, Gas, Bytes) { ... }
    fn create(
        &mut self,
        data: &mut EVMData<'_, DB>,
        inputs: &mut CreateInputs
    ) -> (InstructionResult, Option<Address>, Gas, Bytes) { ... }
    fn create_end(
        &mut self,
        data: &mut EVMData<'_, DB>,
        inputs: &CreateInputs,
        ret: InstructionResult,
        address: Option<Address>,
        remaining_gas: Gas,
        out: Bytes
    ) -> (InstructionResult, Option<Address>, Gas, Bytes) { ... }
    fn selfdestruct(
        &mut self,
        contract: Address,
        target: Address,
        value: Uint<256, 4>
    ) { ... }
}

EVM Interpreter callbacks.

Provided Methods

fn initialize_interp( &mut self, interp: &mut Interpreter<'_>, data: &mut EVMData<'_, DB> )

Called before the interpreter is initialized.

If interp.instruction_result is set to anything other than InstructionResult::Continue then the execution of the interpreter is skipped.

fn step(&mut self, interp: &mut Interpreter<'_>, data: &mut EVMData<'_, DB>)

Called on each step of the interpreter.

Information about the current execution, including the memory, stack and more is available on interp (see Interpreter).

Example

To get the current opcode, use interp.current_opcode().

fn log( &mut self, evm_data: &mut EVMData<'_, DB>, address: &Address, topics: &[FixedBytes<32>], data: &Bytes )

Called when a log is emitted.

fn step_end(&mut self, interp: &mut Interpreter<'_>, data: &mut EVMData<'_, DB>)

Called after step when the instruction has been executed.

Setting interp.instruction_result to anything other than InstructionResult::Continue alters the execution of the interpreter.

fn call( &mut self, data: &mut EVMData<'_, DB>, inputs: &mut CallInputs ) -> (InstructionResult, Gas, Bytes)

Called whenever a call to a contract is about to start.

InstructionResulting anything other than InstructionResult::Continue overrides the result of the call.

fn call_end( &mut self, data: &mut EVMData<'_, DB>, inputs: &CallInputs, remaining_gas: Gas, ret: InstructionResult, out: Bytes ) -> (InstructionResult, Gas, Bytes)

Called when a call to a contract has concluded.

InstructionResulting anything other than the values passed to this function ((ret, remaining_gas, out)) will alter the result of the call.

fn create( &mut self, data: &mut EVMData<'_, DB>, inputs: &mut CreateInputs ) -> (InstructionResult, Option<Address>, Gas, Bytes)

Called when a contract is about to be created.

InstructionResulting anything other than InstructionResult::Continue overrides the result of the creation.

fn create_end( &mut self, data: &mut EVMData<'_, DB>, inputs: &CreateInputs, ret: InstructionResult, address: Option<Address>, remaining_gas: Gas, out: Bytes ) -> (InstructionResult, Option<Address>, Gas, Bytes)

Called when a contract has been created.

InstructionResulting anything other than the values passed to this function ((ret, remaining_gas, address, out)) will alter the result of the create.

fn selfdestruct( &mut self, contract: Address, target: Address, value: Uint<256, 4> )

Called when a contract has been self-destructed with funds transferred to target.