cast::traces::identifier

Trait TraceIdentifier

pub trait TraceIdentifier {
    // Required method
    fn identify_addresses<'a, A>(
        &mut self,
        addresses: A,
    ) -> Vec<AddressIdentity<'_>>
       where A: Iterator<Item = (&'a Address, Option<&'a [u8]>, Option<&'a [u8]>)> + Clone;
}

Trace identifiers figure out what ABIs and labels belong to all the addresses of the trace.

Required Methods

fn identify_addresses<'a, A>( &mut self, addresses: A, ) -> Vec<AddressIdentity<'_>>
where A: Iterator<Item = (&'a Address, Option<&'a [u8]>, Option<&'a [u8]>)> + Clone,

Attempts to identify an address in one or more call traces.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.