cast

Trait InspectorExt

pub trait InspectorExt: for<'a> Inspector<Context<BlockEnv, TxEnv, CfgEnv, &'a mut dyn DatabaseExt<Error = DatabaseError>>> {
    // Provided methods
    fn should_use_create2_factory(
        &mut self,
        _context: &mut Context<BlockEnv, TxEnv, CfgEnv, &mut dyn DatabaseExt<Error = DatabaseError>>,
        _inputs: &CreateInputs,
    ) -> bool { ... }
    fn console_log(&mut self, msg: &str) { ... }
    fn is_odyssey(&self) -> bool { ... }
    fn create2_deployer(&self) -> Address { ... }
}

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

Provided Methods

fn should_use_create2_factory( &mut self, _context: &mut Context<BlockEnv, TxEnv, CfgEnv, &mut dyn DatabaseExt<Error = DatabaseError>>, _inputs: &CreateInputs, ) -> bool

Determines whether the DEFAULT_CREATE2_DEPLOYER should be used for a CREATE2 frame.

If this function returns true, we’ll replace CREATE2 frame with a CALL frame to CREATE2 factory.

fn console_log(&mut self, msg: &str)

Simulates console.log invocation.

fn is_odyssey(&self) -> bool

Returns true if the current network is Odyssey.

fn create2_deployer(&self) -> Address

Returns the CREATE2 deployer address.