Module cast::utils

Modules

sputnik_helpers

Constants

DAPP_JSON

Default Path to where the contract artifacts are stored

LOCAL_RPC_URL

Default local RPC endpoint

Functions

dapp_json_path

The path to where the contract artifacts are stored

etherscan_api_key

Reads the ETHERSCAN_API_KEY env variable

find_artifacts_dir

Determines the artifacts directory to use given the root path to a project’s workspace.

find_contracts_dir

Determines the source directory to use given the root path to a project’s workspace.

find_dapp_json_contract

Tries to extract the Contract in the DAPP_JSON file

find_fave_or_alt_path

Returns the right subpath in a dir

find_git_root_path

find_libs

format_tokens

read_secret

Securely reads a secret from stdin, or proceeds to return a fallback value which was provided in cleartext via CLI or env var

rpc_url

The rpc url to use If the ETH_RPC_URL is not present, it falls back to the default http://127.0.0.1:8545

sputnik_cfg

subscriber

Initializes a tracing Subscriber for logging